AutoBZ.jl

AutoBZ — Module

This Julia package provides routines for multi-dimensional Brillouin zone (BZ) integration for applications of Wannier interpolation to the calculation of density of states and optical conductivity with self energies and spin-orbit coupling. It uses algorithms which automatically compute BZ integrals to a specified error tolerance by resolving smooth yet highly localized integrands.

See AutoBZCore.jl on how to create custom integrands and for more details on the algorithms.

In many-body Green's function methods, BZ integrands are localized at a scale determined by a non-zero, but possibly small, system- and temperature-dependent scattering rate. For example, the single-particle retarded Green's function of an electronic system for frequency $\omega$ and reciprocal space vector $\bm{k}$ with chemical potential $\mu$, Hermitian Hamiltonian matrix $H(\bm{k})$, and self-energy matrix $\Sigma(\omega)$, which is given by

\[G(\omega) = \int_{\text{BZ}} d\bm{k}\ \operatorname{Tr} \left[ (\hbar\omega - H(\bm{k}) - \Sigma(\omega))^{-1} \right]\]

is localized about the manifold defined by $\det(\hbar\omega - H(\bm{k}))=0$ (i.e. the Fermi surface when $\hbar\omega=\mu$) by a scattering rate depending on $\operatorname{Im}\ \Sigma(\omega)$.

Package features

Iterated adaptive integration (IAI) with nested calls to QuadGK.jl

Algorithm with logarithmic complexity for increasingly localized integrands

Irreducible Brillouin zone (IBZ) integration for the cubic lattice

Equispace integration (PTR) as described by Kaye et al. [1]

Automatic algorithm that refines $k$-grid to meet requested error

Support for Wannier-interpolated integrands

User-defined integrands based on Bloch Hamiltonians

Density of states (DOS) calculations

Transport calculations based on TRIQS DFTTools

Calculation of transport function and kinetic coefficients

Option to separate intra-band and inter-band contributions

Parallelized calculations available through batchsolve interface of AutoBZCore.jl

Wannier90-based parsers Hamiltonians (*_hr.dat files) and position operators (*_r.dat files)

Automated interpolation for frequency-dependent self-energy data in text files, using EquiBaryInterp.jl and HChebInterp.jl.

IBZ integration for arbitrary symmetry groups (via an interface to SymmetryReduceBZ.jl)

Supports systems with spin-orbit coupling expressed in the self-energy

To start using the package, see the Workflow and Demos sections.

Notes

If you are an interested Python user, see the Python section

To see a poster showcasing calculations with the library, click this link

Contact the developer

Lorenzo Van Muñoz

1Kaye et al. "Automatic, high-order, and adaptive algorithms for Brillouin zone integration"