Equispace integration

Routines

AutoBZ.equispace_integration — Function

equispace_integration(f, l, npt; pre=nothing, pre_eval=generic_pre_eval, int_eval=generic_int_eval)

Evaluate the integral of f over domain l using an equispace grid of npt points per dimension, optionally using precomputation pre

AutoBZ.automatic_equispace_integration — Function

automatic_equispace_integration(f, a, b; kwargs)
automatic_equispace_integration(f, l::IntegrationLimits; npt1=0, pre1=0, npt2=0, pre2=0, pre_eval=generic_pre_eval, int_eval=generic_int_eval, atol=0.0, rtol=1e-3, npt_update=generic_npt_update, maxevals=typemax(Int64))

Automatically evaluates the integral of f over domain l to within the requested error tolerances atol and rtol. Allows optional precomputed data at two levels of grid refinement npt1, pre1 and npt2, pre2 as well as customizable precomputation with pre_eval and evaluation/summation int_eval. Moreover, a function defining an update strategy for npt can be passed as npt_update.

Customization

AutoBZ.generic_npt_update — Function

generic_npt_update(npt::Integer, f, atol, rtol)

Returns a new npt to try and get another digit of accuracy from PTR. This fallback option is a heuristic, since the scaling of the error is generally problem-dependent.

AutoBZ.generic_pre_eval — Function

generic_pre_eval(f, l, npt)

Precomputes the grid points and weights to use for equispace quadrature of f on the domain l while applying the relevant symmetries to l to reduce the number of evaluation points.

AutoBZ.generic_int_eval — Function

generic_int_eval(f, pre, dvol)

Sums the values of f on the precomputed grid points with corresponding integer weights and also multiplies by the mesh cell volume.