AdaptChebInterp

AutoBZ.AdaptChebInterp.adaptchebinterp — Function

adaptchebinterp(f, a, b; order=4, atol=0, rtol=0, norm=norm, maxevals=typemax(Int), initdiv=1, npoints=10)

Return a piecewise polynomial interpolant of f on the interval $[a,b]$ of degree order that is pointwise accurate to the requested tolerances. The error estimate used by this routine to adaptively refine the panels is to bisect the panel and construct an interpolant on each subpanel and to take the error to be the maximum norm of the difference of the original interpolant and the refined interpolants evaluated on an equispaced grid of npoints(order+1) points. If a panel is accepted, the implementation uses the original interpolant for the panel and discards the refinements. (If unnecessary function evaluations to construct the interpolant should be avoided, see fastadaptchebinterp.) norm is also used set the error tolerance of a panel, which is max(atol, maximum(rtol*norm(f.(p)))) where p are the Chebyshev nodes on the panel.

AutoBZ.AdaptChebInterp.fastadaptchebinterp — Function

fastadaptchebinterp(f, a, b; order=15, atol=0, rtol=0, norm=norm, maxevals=typemax(Int), initdiv=1, abs=abs, ncoeffs=3)

Return a piecewise polynomial interpolant of f on the interval $[a,b]$ of degree order that is pointwise accurate to the requested tolerances. The error estimate used by this routine to adaptively refine the panels is the sum of the abs value of the last ncoeffs Chebyshev coefficients of the interpolant on that panel. norm is used set the error tolerance of a panel, which is max(atol, rtol*maximum(norm, f.(p))) where p are the Chebyshev nodes on the panel. In the case of a non-scalar function, abs may have to be an array norm.