Fourier series

Wannier-interpolated Hamiltonians are represented by Fourier series with a modest number of Fourier coefficients. The goal of this page of documentation is to describe the features, interface, and conventions of Fourier series evaluation as implemented by this library.

Representing Hamiltonians with Fourier series

Fourier series represent functions with sinusoids whose frequencies are integer multiples of a fundamental. In the band theory of solids, the fundamental frequencies for a Hamiltonian correspond to the real-space lattice vectors that generate a Bravais lattice, and so it is best to represent the momentum variable $\vec{k}$ in the basis of the reciprocal lattice. Since it is conventional to construct a reciprocal lattice $\{\vec{b}_j\}$ from a Bravais lattice $\{\vec{a}_i\}$ such that $\vec{b}_j \cdot \vec{a}_i = 2\pi\delta_{ij}$, the Fourier series in this library are defined with a phase factor scaled by $2\pi$. Additionally, any coordinate transformations of $\vec{k}$ from the Cartesian basis to the reciprocal lattice basis only modify Brillouin zone integrals by a multiplicative factor of the absolute value of the determinant of the basis transformation. To find a non-trivial example of representing a Hamiltonian in the reciprocal lattice basis, see the Graphene example with ManyOffsetsFourierSeries.

Interface

AutoBZ.Applications.AbstractFourierSeries — Type

AbstractFourierSeries{N}

A supertype for Fourier series that are periodic maps $\R^N \to V$ where $V$ is any vector space. Typically these can be represented by N-dimensional arrays whose elements belong to the vector space. See the manual section on the AbstractFourierSeries interface.

AutoBZ.Applications.period — Function

period(f::AbstractFourierSeries{N}) where {N}

Return a NTuple{N} whose m-th element corresponds to the period of f along its m-th input dimension. Typically, these values set the units of length for the problem.

AutoBZ.Applications.contract — Function

contract(f::AbstractFourierSeries{N}, x::Number, [dim=N]) where {N}

Return another Fourier series of dimension N-1 by summing over dimension dim of f with the phase factors evaluated at x. If N=1, this function should return an AbstractFourierSeries{0} that stores the evaluated Fourier series, but has no more input dimensions to contract.

The default of dim=N is motivated by preserving memory locality in Julia's column-major array format.

contract(f::AbstractFourierSeries{N}, x::SVector{M}) where {N,M}

Contract the outermost indices M of f in order of last(x) to first(x). If M>N, the default behavior is just to try and contract M indices, which will likely lead to an error.

AutoBZ.Applications.value — Function

value(::AbstractFourierSeries{0})

Return the evaluated Fourier series whose indices have all been contracted. Typically, this value has the same units as the Fourier series coefficients.

Additonally, concrete subtypes of AbstractFourierSeries must have an element type, which they can do by extending Base.eltype with a method. For example, if a type MyFourierSeries <: AbstractFourierSeries always returns ComplexF64 outputs, then the correct eltype method to define would be:

Base.eltype(::Type{MyFourierSeries}) = ComplexF64

The type returned should correspond to the vector space $V$ of the output space of the Fourier series, i.e. the output of value should be of this type. For good performance, the eltype should be a concrete type and should be inferrable.

With the above implemented, several methods which define functors for AbstractFourierSeries allow the user (and integration routines) to evaluate the type like a function with the f(x) syntax.

Types

The concrete types listed below all implement the AbstractFourierSeries interface and should cover most use cases.

AutoBZ.Applications.FourierSeries — Type

FourierSeries(coeffs, period::SVector{N,Float64}) where {N}

Construct a Fourier series whose coefficients are given by the coefficient array array coeffs whose eltype should support addition and scalar multiplication, and whose periodicity on the ith axis is given by period[i]. This type represents the Fourier series

\[f(\vec{x}) = \sum_{\vec{n} \in \mathcal I} C_{\vec{n}} \exp(i2\pi\vec{k}_{\vec{n}}\cdot\overrightarrow{x})\]

where $i = \sqrt{-1}$ is the imaginary unit, $C$ is the array coeffs, $\mathcal I$ is CartesianIndices(C), $\vec{n}$ is a CartesianIndex and $\vec{k}_{\vec{n}}$ is equal to $n_j/p_j$ in the $j$th position with $p_j$ the $j$th element of period. Because of the choice to use Cartesian indices to set the phase factors, typically the indices of coeffs should be specified by using an OffsetArray.

FourierSeries(coeffs::AbstractArray{T,N}, period::Real) where {T,N}

If period is a Real, this constructor will infer the number of input dimensions of the Fourier series from the array dimensionality of the coefficients, and period will become the period of all of the dimensions.

AutoBZ.Applications.FourierSeriesDerivative — Type

FourierSeriesDerivative(f::FourierSeries{N}, a::SVector{N}) where {N}

Represent the differential of Fourier series f by a multi-index a of derivatives, e.g. [1,2,...], whose ith entry represents the order of differentiation on the ith input dimension of f. Mathematically, this means

\[\left( \prod_{j=1}^N \partial_{x_j}^{a_j} \right) f(\vec{x}) = \sum_{\vec{n} \in \mathcal I} \left( \prod_{j=1}^N (i 2\pi k_j)^{a_j} \right) C_{\vec{n}} \exp(i2\pi\vec{k}_{\vec{n}}\cdot\overrightarrow{x})\]

where $\partial_{x_j}^{a_j}$ represents the $a_j$th derivative of $x_j$, $i = \sqrt{-1}$ is the imaginary unit, $C$ is the array coeffs, $\mathcal I$ is CartesianIndices(C), $\vec{n}$ is a CartesianIndex and $\vec{k}_{\vec{n}}$ is equal to $n_j/p_j$ in the $j$th position with $p_j$ the $j$th element of period. Because of the choice to use Cartesian indices to set the phase factors, typically the indices of coeffs should be specified by using an OffsetArray. Also, note that termwise differentiation of the Fourier series results in additional factors of $i2\pi$ which should be anticipated for the use case. Also, note that this type can be used to represent fractional differentiation or integration by suitably choosing the $a_j$s.

This is a 'lazy' representation of the derivative because instead of differentiating by computing all of the Fourier coefficients of the derivative upon constructing the object, the evaluator waits until it contracts the differentiated dimension to evaluate the new coefficients.

AutoBZ.Applications.OffsetFourierSeries — Type

OffsetFourierSeries(f::AbstractFourierSeries{N}, q::SVector{N,Float64}) where {N}

Represent a Fourier series whose argument is offset by the vector $\vec{q}$ and evaluates it as $f(\vec{x}-\vec{q})$.

AutoBZ.Applications.ManyFourierSeries — Type

ManyFourierSeries(fs::AbstractFourierSeries{N}...) where {N}

Represents a tuple of Fourier series of the same dimension and periodicity and contracts them all simultaneously.

AutoBZ.Applications.ManyOffsetsFourierSeries — Type

ManyOffsetsFourierSeries(f, qs..., [origin=true])

Represent a Fourier series evaluated at many different points, and contract them all simultaneously, returning them in the order the qs were passed, i.e. (f(x-qs[1]), f(x-qs[2]), ...) The origin keyword decides whether or not to evaluate $f$ without an offset, and if origin is true, the value of $f$ evaluated without an offset will be returned in the first position of the output.

AutoBZ.Applications.BandEnergyVelocity — Function

BandEnergyVelocity(H::FourierSeries{3})

This constructor takes a Fourier series representing the Hamiltonian and also evaluates the band velocities so that the return value after all the dimensions are contracted is a tuple containing (H, ν₁, ν₂, ν₃). The band velocities are defined by dipole operators $\nu_{\alpha} = \frac{1}{\hbar} \partial_{k_{\alpha}} H$ where $k_{\alpha}$ is one of three input dimensions of $H$ and $\hbar=1$. Note that differentiation by $k$ changes the units to have an additional dimension of length and a factor of $2\pi$, so if $H$ has dimensions of energy, $\nu$ has dimensions of energy times length. The caller is responsible for transforming the units of the velocity (i.e. $\hbar$) if they want other units, which can usually be done as a post-processing step.

Methods

AutoBZ.Applications.contract — Method

contract(f::FourierSeries{N}, x::Number, [dim=N]) where {N}

Contract index dim of the coefficients of f at the spatial point x. The default dim is the outermost dimension to preserve memory locality.

contract(f::FourierSeriesDerivative{N}, x::Number, [dim=N]) where {N}

Contract index dim of the coefficients of f at the spatial point x. The default dim is the outermost dimension to preserve memory locality.