Integrands

User-defined

For integrands that can be evaluated by Wannier interpolation, the following data are necessary to define an integrand:

the integrand evaluator

a Fourier series

additional parameters for the evaluator

Consider implementing custom integrands using the generic template type AutoBZCore.FourierIntegrand that is compatible with all of the adaptive and equispace integration routines. Using this interface will automatically optimize multi-dimensional Fourier series evaluations with the various integration routines.

Pre-defined

Types

AutoBZ.GlocIntegrand — Function

GlocIntegrand(h; Σ, ω, μ=0)

Green's function integrands accepting a self energy Σ that can either be a matrix or a function of ω (see the self energy section of the documentation for examples)

AutoBZ.DiagGlocIntegrand — Function

DiagGlocIntegrand(h; Σ, ω, μ=0)

Green's function integrands accepting a self energy Σ that can either be a matrix or a function of ω (see the self energy section of the documentation for examples)

AutoBZ.TrGlocIntegrand — Function

TrGlocIntegrand(h; Σ, ω, μ=0)

Green's function integrands accepting a self energy Σ that can either be a matrix or a function of ω (see the self energy section of the documentation for examples)

AutoBZ.DOSIntegrand — Function

DOSIntegrand(h; Σ, ω, μ=0)

Green's function integrands accepting a self energy Σ that can either be a matrix or a function of ω (see the self energy section of the documentation for examples)

AutoBZ.TransportFunctionIntegrand — Function

TransportFunctionIntegrand(hv; β, μ=0)

Computes the following integral

\[\D_{\alpha\beta} = \int_{\text{BZ}} dk \operatorname{Tr}[\nu_\alpha(k) A(k,\omega_1) \nu_\beta(k) A(k, \omega_2)]\]

AutoBZ.TransportDistributionIntegrand — Function

TransportDistributionIntegrand(hv; Σ, ω₁, ω₂, μ=0)

A function whose integral over the BZ gives the transport distribution

\[\Gamma_{\alpha\beta}(\omega_1, \omega_2) = \int_{\text{BZ}} dk \operatorname{Tr}[\nu_\alpha(k) A(k,\omega_1) \nu_\beta(k) A(k, \omega_2)]\]

Based on TRIQS. See Integrand for more details.

AutoBZ.KineticCoefficientIntegrand — Function

KineticCoefficientIntegrand(bz, alg::AutoBZAlgorithm, hv::AbstracVelocity; Σ, n, β, Ω, μ=0, abstol, reltol, maxiters)
KineticCoefficientIntegrand(lb, ub, alg, hv::AbstracVelocity; Σ, n, β, Ω, μ=0, abstol, reltol, maxiters)

A function whose integral over the BZ gives the kinetic coefficient. Mathematically, this computes

\[A_{n,\alpha\beta}(\Omega) = \int_{-\infty}^{\infty} d \omega (\beta\omega)^{n} \frac{f(\omega) - f(\omega+\Omega)}{\Omega} \Gamma_{\alpha\beta}(\omega, \omega+\Omega)\]

where $f(\omega) = (e^{\beta\omega}+1)^{-1}$ is the Fermi distriubtion. Based on TRIQS. The argument alg determines what the order of integration is. Given a BZ algorithm, the inner integral is the BZ integral. Otherwise it is the frequency integral.

AutoBZ.OpticalConductivityIntegrand — Function

OpticalConductivityIntegrand

Returns a KineticCoefficientIntegrand with n=0. See KineticCoefficientIntegrand for further details

AutoBZ.ElectronDensityIntegrand — Function

ElectronDensityIntegrand(bz, alg::AutoBZAlgorithm, h::AbstractHamiltonianInterp; Σ, β, [μ=0])
ElectronDensityIntegrand(lb, ub, alg, h::AbstractHamiltonianInterp; Σ, β, [μ=0])

A function whose integral over the BZ gives the electron density. Mathematically, this computes

\[n(\mu) = \int_{-\infty}^{\infty} d \omega f(\omega) \operatorname{DOS}(\omega+\mu)\]

where $f(\omega) = (e^{\beta\omega}+1)^{-1}$ is the Fermi distriubtion. The argument alg determines what the order of integration is. Given a BZ algorithm, the inner integral is the BZ integral. Otherwise it is the frequency integral.

To get the density/number of electrons, multiply the result of this integral by n_sp/det(bz.B)

AutoBZ.AuxTransportDistributionIntegrand — Function

AuxTransportDistributionIntegrand(hv, [auxfun]; Σ, ω₁, ω₂, μ)

A function whose integral over the BZ gives the transport distribution

\[\Gamma_{\alpha\beta}(\omega_1, \omega_2) = \int_{\text{BZ}} dk \operatorname{Tr}[\nu_\alpha(k) A(k,\omega_1) \nu_\beta(k) A(k, \omega_2)]\]

Based on TRIQS. See Integrand for more details.

AutoBZ.AuxKineticCoefficientIntegrand — Function

AuxKineticCoefficientIntegrand

A kinetic coefficient integrand that is more robust to the peak-missing problem. See KineticCoefficientIntegrand for arguments.

AutoBZ.AuxOpticalConductivityIntegrand — Function

AuxOpticalConductivityIntegrand

Returns a AuxKineticCoefficientIntegrand with n=0. See AuxKineticCoefficientIntegrand for further details

Functions

AutoBZ.fermi — Function

fermi(β, ω)
fermi(x)

Evaluates a Fermi distribution with unitless input

\[f(x) = \frac{1}{e^{x}+1}\]

AutoBZ.fermi′ — Function

fermi′(x)

Evaluates a first derivative of the Fermi distribution with unitless input

\[\partial_{x} f(x) = -\frac{1}{2(\cosh(x)+1)}\]

Note that the analytic expression above can be rewritten many ways using hypertrigonometric identities.

AutoBZ.fermi_window — Function

fermi_window(β, ω, Ω)
fermi_window(x, y)

Evaluates a unitless window function with unitless inputs determined by the Fermi distribution $f$ and defined by

\[\chi(x, y) = \frac{f(x) - f(x+y)}{y}\]

In the case y==0 then this simplifies to the derivative of the Fermi distribution.

See also fermi and fermi′.

AutoBZ.fermi_window_limits — Function

fermi_window_limits(Ω, β [; atol=0.0, rtol=1e-20])

Returns limits (a,b) over ω restricted to the interval where the Fermi window is larger than max(atol,rtol*fermi_window(0,β*Ω)). Choosing atol and rtol wisely is important to integrating the entire region of interest, since this is a truncation of an infinite interval, and should be tested for convergence.