Self energies

Interface

AutoBZ.Applications.AbstractSelfEnergy — Type

AbstractSelfEnergy

An abstract type whose instances implement the following interface:

instances are callable and return a square matrix as a function of frequency

instances have methods lb and ub that return the lower and upper bounds of of the frequency domain for which the instance can be evaluated

AutoBZ.Applications.lb — Function

lb(::AbstractSelfEnergy)

Return the greatest lower bound of the domain of the self energy evaluator

AutoBZ.Applications.ub — Function

ub(::AbstractSelfEnergy)

Return the least upper bound of the domain of the self energy evaluator

Types

AutoBZ.Applications.EtaEnergy — Type

EtaEnergy(η::Real)

Construct a self-energy evaluator which returns $-i\eta I$ for any frequency.

AutoBZ.Applications.ScalarEnergy — Type

ScalarEnergy(interpolant, lb, ub)

Construct a self-energy evaluator which for frequencies above lb and below ub returns the interpolant at that frequency times an identity matrix.