EquiBaryInterp

AutoBZ.EquiBaryInterp.LocalEquiBaryInterp — Type

LocalEquiBaryInterp(x::AbstractVector, y::AbstractVector, [degree=8])
LocalEquiBaryInterp(x::Vector, y::Vector, w::Vector, h)

Construct a local barycentric Lagrange interpolant that forms a degree degree local polynomial approximation of the data y on the equispace grid x, which must be identical to a range with step size h. w are the equispace interpolation weights.

AutoBZ.EquiBaryInterp.BaryPoly — Type

BaryPoly(x, y)
BaryPoly(x, y, w)

Constructs a barycentric Lagrange polynomial from the data y sampled on x.

AutoBZ.EquiBaryInterp.barycentric_weights — Function

barycentric_weights(x::AbstractVector{T}) where {T<:Real}

Computes barycentric weights for the nodes x.