Jobs

The module AutoBZ.Jobs has batch functions that setup kinetic coefficient calculations.

Scripts

AutoBZ.Jobs.run_kinetic — Function

run_kinetic(HV, Σ::AbstractSelfEnergy, β, μ, n, Ωs, rtol, atol)

Returns a NamedTuple with names A, err, t, Omega containing the results, errors, and timings for a kinetic coefficient calculation done at frequencies Ωs with parameters β, μ, n, atol, rtol. This function constructs a KineticIntegrand for each parameter value and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results.

AutoBZ.Jobs.run_kinetic_equispace — Function

run_kinetic_equispace(HV, Σ::AbstractSelfEnergy, β, μ, n, Ωs, npt, rtol, atol; pre_eval=pre_eval_contract)

Returns a NamedTuple with names A, err, t, Omega containing the results, errors, and timings for a kinetic coefficient calculation done at frequencies Ωs with parameters β, μ, n, atol, rtol. This function constructs an EquispaceKineticIntegrand for each parameter value, and precomputes HV on an equispace $k$ grid with npt points per dimension (which is reused for all parameter values), and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results.

AutoBZ.Jobs.run_kinetic_auto — Function

run_kinetic_auto(HV, Σ::AbstractSelfEnergy, β, μ, n, Ωs, rtol, atol; ertol=1.0, eatol=0.0)

Returns a NamedTuple with names A, err, t, Omega containing the results, errors, and timings for a kinetic coefficient calculation done at frequencies Ωs with parameters β, μ, n, atol, rtol. This function constructs both an AutoEquispaceKineticIntegrand with wide tolerances eatol and ertol which estimates the integral, int, and then uses a narrow absolute tolerance set by max(atol,rtol*norm(int)) to construct a KineticIntegrand for each parameter value and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results.

Since this is intended to compute a cheap equispace integral first, it is recommended to over-ride the default $k$-grid refinement step to something $\eta$-independent with a line like the one below before calling this script

AutoBZ.equispace_npt_update(npt, ::TransportIntegrand, atol, rtol) = npt + 50

AutoBZ.Jobs.run_kinetic_auto_equispace — Function

run_kinetic_auto_equispace(HV, Σ::AbstractSelfEnergy, β, μ, n, Ωs, rtol, atol)

Returns a NamedTuple with names A, err, t, Omega containing the results, errors, and timings for a kinetic coefficient calculation done at frequencies Ωs with parameters β, μ, n, atol, rtol. This function constructs an AutoEquispaceKineticIntegrand for each parameter value, reusing $k$-grids of HV values from previous calculations, and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results.

Parallelized scripts

AutoBZ.Jobs.run_dos_adaptive_parallel — Function

run_dos_adaptive_parallel(H, Σ::AbstractSelfEnergy, μ, ωs, rtol, atol, [nthreads=Threads.nthreads()])

Returns a NamedTuple with names D, err, t, omega containing the results, errors, and timings for a density of states calculation done at frequencies ωs with parameters μ, atol, rtol. This function constructs a DOSIntegrand for each parameter value and calls iterated_integration on it over the domain of the IBZ to get the results.

AutoBZ.Jobs.run_dos_equispace_parallel — Function

run_dos_equispace_parallel(HV, Σ::AbstractSelfEnergy, μ, ωs, npt, [nthreads=Threads.nthreads()])

Returns a NamedTuple with names D, t, omega containing the results and timings obtained for a density of states calculation done at frequencies ωs with parameters μ, npt, where npt is the number of $k$ points per dimension. The caller should check that the result is converged with respect to npt.

AutoBZ.Jobs.run_dos_auto_equispace_parallel — Function

run_dos_auto_equispace_parallel(HV, Σ::AbstractSelfEnergy, μ, ωs, rtol, atol, [nthreads=1])

Returns a NamedTuple with names D, err, t, omega, npt1, npt2 containing the results, errors, timings, and number of kpts per dimension used for a density of states calculation done at frequencies ωs with parameters μ, atol, rtol.

AutoBZ.Jobs.run_dos_parallel — Function

run_dos_parallel(HV, Σ::AbstractSelfEnergy, μ, ωs, rtol, atol, [nthreads=Threads.nthreads()]; ertol=1.0, eatol=0.0)

Returns a NamedTuple with names D, err, t, omega containing the results, errors, and timings for a density of states calculation done at frequencies ωs with parameters μ, atol, rtol. This function first estimates the integral, int, with an equispace method using wide tolerances eatol and ertol, and then uses a narrow absolute tolerance set by max(atol,rtol*norm(int)) to evaluate the same integral adaptively.

Since this is intended to compute a cheap equispace integral first, it is recommended to over-ride the default $k$-grid refinement step to something $\eta$-independent with a line like the one below before calling this script

AutoBZ.equispace_npt_update(npt, ::TransportIntegrand, atol, rtol) = npt + 50

AutoBZ.Jobs.run_kinetic_parallel — Function

run_kinetic_parallel(filename, HV, Σ::AbstractSelfEnergy, β, μ, n, Ωs, rtol, atol; nthreads=Threads.nthreads())

Writes an h5 archive to filename with groups A, err, t, Omega containing the results, errors, and timings for a kinetic coefficient calculation done at frequencies Ωs with parameters β, μ, n, atol, rtol. This function constructs a KineticIntegrand for each parameter value and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results. The calculation is parallelized over Ωs on nthreads threads.

AutoBZ.Jobs.run_kinetic_equispace_parallel — Function

run_kinetic_equispace_parallel(filename, HV, Σ::AbstractSelfEnergy, β, μ, n, Ωs, npt, rtol, atol; pre_eval=pre_eval_contract, nthreads=Threads.nthreads())

Writes an h5 archive to filename with groups A, err, t, Omega containing the results, errors, and timings for a kinetic coefficient calculation done at frequencies Ωs with parameters β, μ, n, atol, rtol. This function constructs an EquispaceKineticIntegrand for each parameter value, and precomputes HV on an equispace $k$ grid with npt points per dimension (which is reused for all parameter values), and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results. The calculation is parallelized over Ωs on nthreads threads.

AutoBZ.Jobs.run_kinetic_auto_parallel — Function

run_kinetic_auto_parallel(HV, Σ::AbstractSelfEnergy, β, μ, n, Ωs, rtol, atol; ertol=1.0, eatol=0.0, nthreads=Threads.nthreads())

Returns a NamedTuple with names A, err, t, Omega containing the results, errors, and timings for a kinetic coefficient calculation done at frequencies Ωs with parameters β, μ, n, atol, rtol. This function constructs both an AutoEquispaceKineticIntegrand with wide tolerances eatol and ertol which estimates the integral, int, and then use a narrow absolute tolerance set by max(atol,rtol*norm(int)) to construct a KineticIntegrand for each parameter value and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results. The calculation is parallelized over Ωs on nthreads threads.

Since this is intended to compute a cheap equispace integral first, it is recommended to over-ride the default $k$-grid refinement step to something $\eta$-independent with a line like the one below before calling this script

AutoBZ.equispace_npt_update(npt, ::TransportIntegrand, atol, rtol) = npt + 50

AutoBZ.Jobs.run_kinetic_auto_equispace_parallel — Function

run_kinetic_auto_equispace_parallel(filename, HV, Σ::AbstractSelfEnergy, β, μ, n, Ωs, rtol, atol; nthreads=1)

Writes an h5 archive to filename with groups A, err, t, Omega containing the results, errors, and timings for a kinetic coefficient calculation done at frequencies Ωs with parameters β, μ, n, atol, rtol. This function constructs an AutoEquispaceKineticIntegrand for each parameter value, reusing $k$-grids of HV values from previous calculations, and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results. The calculation is parallelized over Ωs on nthreads threads. The default is set to 1 thread for frequency parallelization, although k-point parallelization is still enabled, to avoid duplicating calculations of the Hamiltonian and band velocities on the k-mesh.

I/O

AutoBZ.Jobs.read_h5_to_nt — Function

read_h5_to_nt(filename)

Loads the h5 archive from filename and reads its datasets into a NamedTuple

AutoBZ.Jobs.write_nt_to_h5 — Function

write_nt_to_h5(nt::NamedTuple, filename)

Takes a NamedTuple and writes its values, which must be arrays, into an h5 archive at filename with dataset names corresponding to the tuple names.

AutoBZ.Jobs.import_self_energy — Function

import_self_energy(filename)

Reads the groups omega and sigma in the h5 archive in filename and tries save it to a NamedTuple with names ω and Σ. The array in sigma should be of size (length(omega), 2), where the two columns are the real and imaginary parts of Σ.

Internal

AutoBZ.Jobs.get_safe_freq_limits — Function

get_safe_freq_limits(Ωs, β, lb, ub)

Given a collection of frequencies, Ωs, returns a Vector{CubicLimits{1}} with truncated limits of integration for the frequency integral at each (Ω, β) point that are determined by the fermi_window_limits routine set to the default tolerances for the decay of the Fermi window function. The arguments lb and ub are lower and upper limits on the frequency to which the default result gets truncated if the default result would recommend a wider interval. If there is any truncation, a warning is emitted to the user, but the program will continue with the truncated limits.

AutoBZ.Jobs.batch_smooth_param — Function

batch_smooth_param(xs, nthreads)

If the cost of a calculation smoothly varies with the parameters xs, then batch xs into nthreads groups where the ith element of group j is xs[j+(i-1)*nthreads]