Adaptive integration

Routines

AutoBZ.tree_integration — Function

tree_integration(f, a, b)
tree_integration(f, ::CubicLimits)

Calls HCubature to perform multi-dimensional integration of f over a cube.

AutoBZ.iterated_integration — Function

iterated_integration(f, ::IntegrationLimits)

Calls HCubature to perform iterated 1D integration of f over a domain parametrized by IntegrationLimits.

Customization

The methods below can be extended/customized to integrand types

AutoBZ.iterated_tol_update — Function

iterated_tol_update(f, l, atol, rtol)

Choose a new set of error tolerances for the next inner integral. By default returns (atol, rtol) unchanged.

AutoBZ.iterated_pre_eval — Function

iterated_pre_eval(f, x)

Perform a precomputation

AutoBZ.thunk — Function

thunk(f, x)

Delay the computation of f(x). Needed to normally evaluate an integrand in nested integrals as employed by callbacks.

AutoBZ.ThunkIntegrand — Type

ThunkIntegrand(f, x)

Store f and x to evaluate f(x) at a later time. Can be employed in iterated integration