Equispace integration

Routines

AutoBZ.equispace_integration — Function

equispace_integration(f, l, npt; pre=nothing, equispace_pre_eval=generic_equispace_pre_eval, equispace_int_eval=generic_equispace_int_eval)

Evaluate the integral of f over domain l using an equispace grid of npt points per dimension, optionally using precomputation pre

AutoBZ.automatic_equispace_integration — Function

automatic_equispace_integration(f, a, b; kwargs)
automatic_equispace_integration(f, l::IntegrationLimits; npt1=0, pre1=0, npt2=0, pre2=0, atol=0.0, rtol=1e-3 maxevals=typemax(Int64))

Automatically evaluates the integral of f over domain l to within the requested error tolerances atol and rtol. Allows optional precomputed data at two levels of grid refinement npt1, pre1 and npt2, pre2.

Customization

The methods below can be extended/customized to integrand types

AutoBZ.equispace_npt_update — Function

equispace_npt_update(npt, f, atol, rtol)

Returns a new npt to try and get another digit of accuracy from PTR. This fallback option is a heuristic, since the scaling of the error is generally problem-dependent.

AutoBZ.equispace_pre_eval — Function

equispace_pre_eval(f, l, npt)

Precomputes the grid points and weights to use for equispace quadrature of f on the domain l while applying the relevant symmetries to l to reduce the number of evaluation points. Should return a vector of tuples with the integration weight in the first position and the precomputation in the second. This output is passed to the argument pre of equispace_int_eval.

equispace_pre_eval(f::WannierIntegrand, l::IntegrationLimits, npt)

This function will evaluate the Fourier series and integration weights needed for equispace integration of f at npt points per dimension. l should contain the relevant symmetries needed for IBZ integration, if desired.

AutoBZ.equispace_int_eval — Function

equispace_int_eval(f, pre, dvol)

Sums the values of f on the precomputed grid points with corresponding quadrature weights and also multiplies by the mesh cell volume to obtain the integral on the precomputed domain represented by pre, obtained from equispace_pre_eval. Evaluation of f is done by the function evaluate_integrand in order to create a function boundary between the quadrature and the integrand evaluation.

AutoBZ.evaluate_integrand — Function

evaluate_integrand(f, x)

By default, this calls f(x), however the caller may dispatch on the type of f if they would like to specialize this function together with equispace_pre_eval so that x is a more useful precomputation (e.g. a Fourier series evaluated at a grid point).