Fourier series

Wannier-interpolated Hamiltonians are represented by Fourier series with a modest number of Fourier coefficients. The goal of this page of documentation is to describe the features, interface, and conventions of Fourier series evaluation as implemented by this library.

Conventions

Fourier series represent functions as linear combinations of sinusoids whose frequencies are integer multiples of a fundamental mode. In the band theory of solids, the fundamental frequencies, or normal modes, for a Hamiltonian correspond to linear combinations of real-space lattice vectors that generate a Bravais lattice. The sections below define conventions for each of these linear combinations.

Lattice vectors

It is conventional to construct a reciprocal lattice $\{\bm{b}_j\}$ from a Bravais lattice $\{\bm{a}_i\}$ such that

\[\bm{b}_j \cdot \bm{a}_i = 2\pi\delta_{ij}\]

Then we write the momentum space variable in the reciprocal lattice vector basis and the position space variable in the real lattice vector basis

\[\bm{k} = \sum_{j=1}^{d} k_j \bm{b}_j \qquad \bm{R} = \sum_{i=1}^{d} R_i \bm{a}_i\]

Without loss of generality, the $k_j$ can be taken in the domain $[0,1]$ and the $R_i$ are integers.

Additionally, any coordinate transformations of $\bm{k}$ from the Cartesian basis to the reciprocal lattice basis only modify Brillouin zone integrals by a multiplicative factor of the absolute value of the determinant of the basis transformation.

Series coefficients

Wannier-interpolated Hamiltonians are small matrices obtained by projecting the ab-initio Hamiltonian onto a low energy subspace, a process called downfolding. These Hamiltonians can be expressed by a Fourier series as in the sum below

\[H(\bm{k}) = \sum_{\bm{R}} e^{i\bm{k}\cdot\bm{R}} H_{\bm{R}}\]

where the coefficients $H_{\bm{R}}$ are the matrix-valued Fourier coefficients. Truncating the sum over $\bm{R}$ at a modest number of modes can be done for Wannier Hamiltonians in the maximally-localized orbital basis, for which $H(\bm{k})$ is a smooth and periodic function and thus the truncation error of its Fourier series converges super-algebraically with respect to the number of modes.

Hamiltonian recipe

In model systems, a Bloch Hamiltonian can often be written down analytically. The recipe to write it as a Fourier series has two-steps

Identify the real and reciprocal Bravais lattices, $\{\bm{a}_i\}$ and $\{\bm{b}_j\}$, and rewrite all of the phase dependences of the Hamiltonian as $\bm{k}\cdot\bm{R}$ with each vector in its corresponding basis, as explained above.

Factor the Hamiltonian into a linear combination of normal modes indexed by the distinct $\bm{R}$ vectors. If the Hamiltonian is matrix-valued, this can be done one matrix element at a time.

Interface

See FourierSeriesEvaluators.jl for the AbstractFourierSeries interface, which allows evaluation of the series with a function-like f(x) syntax

Types

The concrete types listed below all implement the AbstractFourierSeries interface and should cover most use cases.

AutoBZ.AbstractWannierInterp — Type

AbstractWannierInterp{N,T,iip} <: AbstractFourierSeries{N,T,iip}

Abstract supertype for all Wannier-interpolated quantities in AutoBZ

AutoBZ.AbstractGaugeInterp — Type

AbstractGaugeInterp{G,N,T,iip} <: AbstractWannierInterp{N,T,iip}

An abstract subtype of AbstractFourierSeries representing in-place Fourier series evaluators for Wannier-interpolated quantities with a choice of basis, or gauge, G, which is typically Hamiltonian or Wannier. For details, see to_gauge.

AutoBZ.AbstractHamiltonianInterp — Type

AbstractHamiltonianInterp{G,N,T,iip} <: AbstractGaugeInterp{G,N,T,iip}

Abstract type representing Hamiltonians, which are matrix-valued Hermitian Fourier series. They should also have period 1, but produce derivatives with wavenumber 1.

AutoBZ.HamiltonianInterp — Type

HamiltonianInterp(f::FourierSeries; gauge=:Wannier)

A wrapper for FourierSeries with an additional gauge that allows for convenient diagonalization of the result. For details see to_gauge.

AutoBZ.SOCHamiltonianInterp — Type

SOCHamiltonianInterp(f::Freq2RadSeries, λ; gauge=Wannier())

A wrapper for a Fourier series in a given gauge that has spin-orbit coupling represented by the matrix λ. In particular, this interpolates [f(k) 0; 0 f(k)] + λ.

AutoBZ.AbstractCoordInterp — Type

AbstractCoordInterp{B,G,N,T,iip} <:AbstractGaugeInterp{G,N,T,iip}

An abstract subtype of AbstractGaugeInterp also containing information about the coordinate basis B, which is either Lattice or Cartesian. For details see to_coord and CoordDefault.

AutoBZ.BerryConnectionInterp — Type

BerryConnectionInterp{P}(a::ManyFourierSeries, B; coord)

Interpolate the Berry connection in basis coord. a must evaluate the components of the connection in coordinate basis P, and B is the coordinate transformation from P to coord.

AutoBZ.AbstractVelocityInterp — Type

AbstractVelocityInterp{C,B,G,N,T,iip} <:AbstractCoordInterp{B,G,N,T,iip}

An abstract subtype of AbstractCoordInterp also containing information the velocity component, C, which is typically Val(:whole), Val(:inter), or Val(:intra). For details see to_vcomp_gauge. Since the velocity depends on the Hamiltonian, subtypes should also evaluate the Hamiltonian.

AutoBZ.GradientVelocityInterp — Type

GradientVelocityInterp(h::AbstractHamiltonianInterp, A; gauge, coord, vcomp)

Evaluate the Hamiltonian and its gradient, which doesn't produce gauge-covariant velocities. The Hamiltonian h must be in the Wannier gauge, but this will give the result in the requested gauge. A must be the coordinate transformation from the lattice basis to the desired coord system. vcomp selects the contribution to the velocities.

AutoBZ.CovariantVelocityInterp — Type

CovariantVelocityInterp(hv::GradientVelocityInterp, a::BerryConnectionInterp)

Uses the Berry connection to return fully gauge-covariant velocities. Returns a tuple of the Hamiltonian and the three velocity matrices.

AutoBZ.MassVelocityInterp — Type

MassVelocityInterp(h::AbstractHamiltonianInterp, A; gauge, coord, vcomp)

Compute the Hamiltonian, its gradient and Hessian, which are not gauge-covariant. See GradientVelocityInterp for explanation of the arguments

Gauges

AutoBZ.AbstractGauge — Type

AbstractGauge

Abstract supertype of gauges (or bases) for orbital/band indices

AutoBZ.Wannier — Type

Wannier <: AbstractGauge

Singleton type representing the Wannier gauge

AutoBZ.Hamiltonian — Type

Hamiltonian <: AbstractGauge

Singleton type representing the Hamiltonian gauge

Coordinate systems

AutoBZ.AbstractCoordinate — Type

AbstractCoordinate

Abstract supertype of bases for coordinate/spatial indices

AutoBZ.Cartesian — Type

Cartesian <: AbstractCoordinate

Singleton type representing Cartesian coordinates.

AutoBZ.Lattice — Type

Lattice <: AbstractCoordinate

Singleton type representing lattice coordinates.

Velocity components

AutoBZ.AbstractVelocityComponent — Type

AbstractVelocityComponent

Abstract supertype representing velocity components.

AutoBZ.Whole — Type

Whole <: AbstractVelocityComponent

Singleton type representing whole velocities

AutoBZ.Intra — Type

Intra <: AbstractVelocityComponent

Singleton type representing intraband velocities

AutoBZ.Inter — Type

Inter <: AbstractVelocityComponent

Singleton type representing interband velocities

Special matrix types

AutoBZ.SOCMatrix — Type

SOCMatrix(A)

Wrapper for a matrix A that should be used as a block-diagonal matrix [A 0 0 A] as when dealing with spin-orbit coupling

AutoBZ.SSymmetricCompact — Type

SSymmetricCompact{N, T, L} <: StaticMatrix{N, N, T}

A StaticArray subtype that represents a Symmetric matrix. Unlike LinearAlgebra.Symmetric, SSymmetricCompact stores only the lower triangle of the matrix (as an SVector). The lower triangle is stored in column-major order. For example, for an SSymmetricCompact{3}, the indices of the stored elements can be visualized as follows:

┌ 1 ⋅ ⋅ ┐
| 2 4 ⋅ |
└ 3 5 6 ┘

Type parameters:

N: matrix dimension;

T: element type for lower triangle;

L: length of the SVector storing the lower triangular elements.

Note that L is always the Nth triangular number.

An SSymmetricCompact may be constructed either:

from an AbstractVector containing the lower triangular elements; or

from a Tuple containing both upper and lower triangular elements in column major order; or

from another StaticMatrix.

For the latter two cases, only the lower triangular elements are used; the upper triangular elements are ignored.