Integration limits

Interface

AutoBZ.IntegrationLimits — Type

IntegrationLimits{d}

Represents a set of integration limits over d variables. Realizations of this type should implement lower and upper, which return the lower and upper limits of integration along some dimension, rescale which represents the number of symmetries of the BZ which are used by the realization to reduce the BZ (the integrand over the limits gets multiplied by this factor), and a functor that accepts a single numeric argument and returns another realization of that type (in order to do nested integration). Thus the realization is also in control of the order of variables of integration and must coordinate this behavior with their integrand. Instances should also be static structs.

AutoBZ.lower — Function

lower(::IntegrationLimits)

Return the lower limit of the next variable of integration. If a vector is returned, then the integration routine may attempt to multidimensional integration.

AutoBZ.upper — Function

upper(::IntegrationLimits)

Return the upper limit of the next variable of integration. If a vector is returned, then the integration routine may attempt to multidimensional integration.

AutoBZ.box — Function

box(::IntegrationLimits)

Return an iterator of tuples that for each dimension returns a tuple with the lower and upper limits of the integration domain without symmetries applied.

AutoBZ.symmetries — Function

symmetries(::IntegrationLimits)

Return an iterator over the symmetry transformations that the parametrization has used to reduce the volume of the integration domain.

Base.ndims — Function

ndims(::IntegrationLimits{d})

Returns d. This is a type-based rule.

AutoBZ.nsyms — Function

nsyms(::IntegrationLimits)

Return the number of symmetries that the parametrization has used to reduce the volume of the integration domain.

Additionally, all IntegrationLimits must extend Base.eltype to return the type which is the output of lower and upper, which is the type of coordinates in the domain.

Types

AutoBZ.CubicLimits — Type

CubicLimits(a, [b])

Store integration limit information for a hypercube with vertices a and b. If b is not passed as an argument, then the lower limit defaults to zero(a).

AutoBZ.CompositeLimits — Type

CompositeLimits(lims::IntegrationLimits...)
CompositeLimits(::Tuple{Vararg{IntegrationLimits}})

Construct a collection of limits which yields the first limit followed by the second, and so on.

Routines

AutoBZ.vol — Function

vol(::IntegrationLimits)

Return the volume of the full domain without the symmetries applied

AutoBZ.symmetrize — Function

symmetrize(::IntegrationLimits, x)
symmetrize(::IntegrationLimits, xs...)

Transform x by the symmetries of the parametrization used to reduce the domain, thus mapping the value of x on the parametrization to the full domain. When the integrand is a scalar, this is equal to nsyms(l)*x. When the integrand is a vector, this is sum(S*x for S in symmetries(l)). When the integrand is a matrix, this is sum(S*x*S' for S in symmetries(l)).

AutoBZ.discretize_equispace — Function

discretize_equispace(::IntegrationLimits, ::Integer)

Return an iterator of 2-tuples containing integration nodes and weights that correspond to an equispace integration grid with the symmetry transformations applied to it.