Interfaces

Wannier90

AutoBZ.load_wannier90_data — Function

load_wannier90_data(seedname; velocity_kind=:none, read_pos_op=true, compact=:N)

Reads Wannier90 output files with the given seedname to return the Hamiltonian (optionally with band velocities if velocity_kind is specified, and orbital velocities if read_pos_op is true) and the full Brillouin zone limits. The keyword compact is available if to compress the Fourier series if its Fourier coefficients are known to be Hermitian.

AutoBZ.load_hamiltonian — Function

load_hamiltonian(filename; period=1.0, compact=:N)

Load an ab-initio Hamiltonian output from Wannier90 into filename as an evaluatable FourierSeries whose periodicity can be set by the keyword argument period which defaults to setting the period along each dimension to 1.0. To define different periods for different dimensions, pass an SVector as the period. To store Hermitian Fourier coefficients in compact form, use the keyword compact to specify:

:N: do not store the coefficients in compact form

:L: store the lower triangle of the coefficients

:U: store the upper triangle of the coefficients

:S: store the lower triangle of the symmetrized coefficients, (c+c')/2

AutoBZ.load_hamiltonian_velocities — Function

load_hamiltonian_velocities(f_hamiltonian, [f_pos_op]; period=1.0, compact=:N, kind=:full)

Load the Hamiltonian and band velocities, which may later be passed to one of the integrand constructors. When called with one filename, that file is parsed as a Wannier 90 Hamiltonian and the resulting Band velocities are just the gradient of the Hamiltonian. The return type is BandEnergyVelocity3D. When called with two filenames, the second is parsed as a position operator from Wannier 90 and adds a contribution to band velocities from the Berry connection. The return type is BandEnergyBerryVelocity3D.The keywords period and compact set the reciprocal unit cell length and whether the coefficients of the Fourier series should be compressed as Hermitian matrices. Typically the coefficients cannot be compressed despite the values of the series being Hermitian. The keyword kind can take values :full, :inter, and :intra, referring to whether the Hamiltonian and band velocities are kept in the Wannier gauge, or rotated to the Hamiltonian gauge to take just the inter/intra-band part (i.e. the off-diagonal/diagonal part of the band velocities, respectively).

Warning: a kind of :inter or :intra is incompatible with a non-scalar Σ

AutoBZ.load_position_operator — Function

load_position_operator(filename; period=1.0, compact=nothing)

Load a position operator Hamiltonian output from Wannier90 into filename as an evaluatable ManyFourierSeries with separate x, y, and z components whose periodicity can be set by the keyword argument period which defaults to setting the period along each dimension to 1.0. To define different periods for different dimensions, pass an SVector as the period. To store Hermitian Fourier coefficients in compact form, use the keyword compact to specify:

:N: do not store the coefficients in compact form

:L: store the lower triangle of the coefficients

:U: store the upper triangle of the coefficients

:S: store the lower triangle of the symmetrized coefficients, (c+c')/2

Note that in some cases the coefficients are not Hermitian even though the values of the series are.

AutoBZ.parse_hamiltonian — Function

parse_hamiltonian(filename)

Parse an ab-initio Hamiltonian output from Wannier90 into filename, extracting the fields (date_time, num_wann, nrpts, degen, irvec, C)

AutoBZ.parse_position_operator — Function

parse_position_operator(filename)

Parse a position operator output from Wannier90 into filename, extracting the fields (date_time, num_wann, nrpts, irvec, X, Y, Z)

Python

Julia code, including AutoBZ.jl, can be called from Python using the package PyJulia

Setup

TL;DR

$ julia -e 'import Pkg; Pkg.add("PyCall")'
$ python3 -m pip install julia

If you want to, you can install PyJulia in a Python venv, but on the Julia side PyCall.jl must be installed in the default environment

Demo

The Python snippet below shows how from the demos folder of the AutoBZ repository you can run one of the demo scripts:

from julia.api import Julia
jl = Julia(compiled_modules=False)

# julia environment setup in working directory 'demos'
jl.eval("""
import Pkg
Pkg.activate(".")
Pkg.instantiate()
""")

# capture output of script
out = jl.eval('include("DOS_test.jl")')

The first two lines are adapted for loading PyJulia on Debian systems.