Integrands

Generic

Types

AutoBZ.Applications.WannierIntegrand — Type

WannierIntegrand(f, s::AbstractFourierSeries, p)

A type generically representing an integrand f whose entire dependence on the variables of integration is in a Fourier series s, and which may also accept some input parameters p, which are preferrably contained in a tuple. The caller must be aware that their function, f, will be called at many evaluation points, x, in the following way: f(s(x), p...). Therefore the caller is expected to know the type of s(x) (hint: eltype(s)) and the layout of the parameters in the tuple p.

Methods

Missing docstring.

Missing docstring for AutoBZ.Applications.pre_eval_contract(::AutoBZ.Applications.WannierIntegrand,::Any,::Any). Check Documenter's build log for details.

Pre-defined

Types

AutoBZ.Applications.DOSIntegrand — Type

DOSIntegrand(H, ω, Σ, μ)
DOSIntegrand(H, ω, Σ)
DOSIntegrand(H, M)

A type whose integral gives the density of states.

\[D(ω) = -{\pi}^{-1} \int_{\text{BZ}} dk \operatorname{Tr}[\Im[{((\omega + \mu) I - H(k) - \Sigma(\omega))}^{-1}]]\]

This type works with both adaptive and equispace integration routines.

AutoBZ.Applications.GammaIntegrand — Type

GammaIntegrand(H::FourierSeries, Σ, ω, Ω, μ, [kind=:full])
GammaIntegrand(HV, Σ, ω, Ω, μ)
GammaIntegrand(HV, Mω, MΩ)

A type whose integral over the BZ gives the transport distribution.

\[\Gamma_{\alpha\beta}(\omega, \Omega) = \int_{\text{BZ}} dk \operatorname{Tr}[\nu_\alpha(k) A(k,\omega) \nu_\beta(k) A(k, \omega+\Omega)]\]

This type works with both adaptive and equispace integration routines. The keyword kind determines the band velocity component.

AutoBZ.Applications.OCIntegrand — Type

OCIntegrand(H::FourierSeries, Σ, Ω, β, μ, [kind=:full])
OCIntegrand(HV, Σ, Ω, β, μ)

A function whose integral over the BZ and the frequency axis gives the optical conductivity. Mathematically, this computes

\[\sigma_{\alpha\beta}(\Omega) = \int_{-\infty}^{\infty} d \omega \frac{f(\omega) - f(\omega+\Omega)}{\Omega} \Gamma_{\alpha\beta}(\omega, \omega+\Omega)\]

where $f(\omega) = (e^{\beta\omega}+1)^{-1}$ is the Fermi distriubtion. Use this type only for adaptive integration and order the limits so that the integral over the Brillouin zone is the outer integral and the frequency integral is the inner integral. The keyword kind determines the band velocity component.

AutoBZ.Applications.EquispaceOCIntegrand — Type

EquispaceOCIntegrand(σ::OCIntegrand, l, npt, pre::Vector{Tuple{NTuple{4, SMatrix{3, 3, ComplexF64, 9}},Int}})
EquispaceOCIntegrand(σ, l, npt)

This type represents an OCIntegrand, σ integrated adaptively in frequency and with equispace integration over the Brillouin zone with a fixed number of grid points npt. The argument l should be an IntegrationLimits for just the Brillouin zone. This type should be called by an adaptive integration routine whose limits of integration are only the frequency variable.

AutoBZ.Applications.AutoEquispaceOCIntegrand — Type

AutoEquispaceOCIntegrand(σ, l, atol, rtol, npt1, pre1, npt2, pre2)
AutoEquispaceOCIntegrand(σ, l, atol, rtol; npt1=0, pre1=Tuple{eltype(σ.HV),Int}[], npt2=0,pre2=Tuple{eltype(σ.HV),Int}[])

This type represents an OCIntegrand, σ integrated adaptively in frequency and with equispace integration over the Brillouin zone with a number of grid points necessary to meet the maximum of the tolerances given by atol and rtol. The argument l should be an IntegrationLimits for just the Brillouin zone. This type should be called by an adaptive integration routine whose limits of integration are only the frequency variable.

The keyword arguments, which are just passed to automatic_equispace_integration, are:

pre1: a Vector containing tuples of the evaluated Hamiltonian + band velocities and integration weights

npt1: an integer that should be equivalent to length(pre1)

pre2: a Vector containing tuples of the evaluated Hamiltonian + band velocities and integration weights on a more refined grid than pre1

npt2: an integer that should be equivalent to length(pre)

Functions

Missing docstring.

Missing docstring for AutoBZ.Applications.dos_integrand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.gamma_integrand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.oc_integrand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.fermi. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.fermi′. Check Documenter's build log for details.

AutoBZ.Applications.fermi_window — Function

Evaluates a unitless window function determined by the Fermi distribution

Missing docstring.

Missing docstring for AutoBZ.Applications.cosh_ratio. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.EXP_P1_SMALL_X. Check Documenter's build log for details.