Integrands

Generic

Types

AutoBZ.Applications.WannierIntegrand — Type

WannierIntegrand(f, s::AbstractFourierSeries, p)

A type generically representing an integrand f whose entire dependence on the variables of integration is in a Fourier series s, and which may also accept some input parameters p, which are preferrably contained in a tuple. The caller must be aware that their function, f, will be called at many evaluation points, x, in the following way: f(s(x), p...). Therefore the caller is expected to know the type of s(x) (hint: eltype(s)) and the layout of the parameters in the tuple p.

Methods

AutoBZ.Applications.pre_eval_contract — Method

pre_eval_contract(f::WannierIntegrand, l::IntegrationLimits, npt)

This function will evaluate the Fourier series and integration weights needed for equispace integration of f at npt points per dimension. l should contain the relevant symmetries needed for IBZ integration, if desired.

Pre-defined

Types

AutoBZ.Applications.GreensFunction — Type

GreensFunction(H,ω,Σ,μ)
GreensFunction(H,ω,Σ)
GreensFunction(H,M)

A struct that calculates the lattice Green's function from a Hamiltonian.

\[G(\omega) = \int_{\text{BZ}} dk {((\omega + \mu) I - H(k) - \Sigma(\omega))}^{-1}\]

AutoBZ.Applications.SpectralFunction — Type

SpectralFunction(::GreensFunction)

A type that whose integral gives the imaginary part of the Green's function.

\[A(ω) = {\pi}^{-1} \Im[G(ω)]\]

AutoBZ.Applications.DOSIntegrand — Type

DOSIntegrand(::SpectralFunction)

A type whose integral gives the density of states.

\[D(ω) = \operatorname{Tr}[A(ω)]\]

AutoBZ.Applications.GammaIntegrand — Type

GammaIntegrand(H, Σ, ω, Ω, μ)
GammaIntegrand(H, ν₁, ν₂, ν₃, Mω, MΩ)

A type whose integral over the BZ gives the transport distribution.

\[\Gamma_{\alpha\beta}(\omega, \Omega) = \int_{\text{BZ}} dk \operatorname{Tr}[\nu_\alpha(k) A(k,\omega) \nu_\beta(k) A(k, \omega+\Omega)]\]

AutoBZ.Applications.OCIntegrand — Type

OCIntegrand(H, ν₁, ν₂, ν₃, Ω, β, η, μ)

A function whose integral over the BZ and the frequency axis gives the optical conductivity. Mathematically, this computes

\[\sigma_{\alpha\beta}(\Omega) = \int_{-\infty}^{\infty} d \omega \Gamma_{\alpha\beta}(\omega, \omega+\Omega)\]

AutoBZ.Applications.EquispaceOCIntegrand — Type

EquispaceOCIntegrand(σ::OCIntegrand, l, npt, pre::Vector{Tuple{NTuple{4, SMatrix{3, 3, ComplexF64, 9}},Int}})
EquispaceOCIntegrand(σ, l, npt, [pre_eval=pre_eval_contract])

This type represents an OCIntegrand, σ integrated adaptively in frequency and with equispace integration over the Brillouin zone with a fixed number of grid points npt. The argument l should be an IntegrationLimits for just the Brillouin zone. This type should be called by an adaptive integration routine whose limits of integration are only the frequency variable.

AutoBZ.Applications.AutoEquispaceOCIntegrand — Type

AutoEquispaceOCIntegrand(σ, l, atol, rtol, pre_eval, npt_update, npt1, pre1, npt2, pre2)
AutoEquispaceOCIntegrand(σ, l, atol, rtol; pre_eval=pre_eval_contract, npt_update=npt_update_sigma, npt1=0, pre1=Tuple{NTuple{4, SMatrix{3, 3, ComplexF64, 9}},Int}[], npt2=0,pre2=Tuple{NTuple{4, SMatrix{3, 3, ComplexF64, 9}},Int}[])

This type represents an OCIntegrand, σ integrated adaptively in frequency and with equispace integration over the Brillouin zone with a number of grid points necessary to meet the maximum of the tolerances given by atol and rtol. The argument l should be an IntegrationLimits for just the Brillouin zone. This type should be called by an adaptive integration routine whose limits of integration are only the frequency variable.

The keyword arguments, which are just passed to automatic_equispace_integration, fall into two categories:

Integrand evaluators: pre_eval and npt_update

Stored precomputations

pre1: a Vector containing tuples of the evaluated Hamiltonian + band velocities and integration weights

npt1: an integer that should be equivalent to length(pre1)

pre2: a Vector containing tuples of the evaluated Hamiltonian + band velocities and integration weights on a more refined grid than pre1

npt2: an integer that should be equivalent to length(pre)

Functions

Missing docstring.

Missing docstring for AutoBZ.Applications.greens_function. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.spectral_function. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.dos_integrand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.gamma_integrand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.oc_integrand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.fermi. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.fermi′. Check Documenter's build log for details.

AutoBZ.Applications.fermi_window — Function

Evaluates a unitless window function determined by the Fermi distribution

Missing docstring.

Missing docstring for AutoBZ.Applications.cosh_ratio. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.EXP_P1_SMALL_X. Check Documenter's build log for details.