Integrands

User-defined

For integrands that can be evaluated by Wannier interpolation, the following data are necessary to define an integrand:

the integrand evaluator

a Fourier series

additional parameters for the evaluator

Consider implementing custom integrands using the generic template type AutoBZ.Applications.WannierIntegrand that is compatible with all of the adaptive and equispace integration routines.

Optimizing equispace integration

Unlike for adaptive integration, the caller is responsible for passing pre-computed grid values to the equispace integration routines, which is explained in the documentation for Equispace integration and AutoBZ.Applications.equispace_pre_eval.

Mixing adaptive and equispace integrals

While it is possible to perform an integral where some variables are integrated adaptively and others are integrated uniformly, this guide will not explain how to do this. However, an example implementation of this is AutoBZ.Applications.AutoEquispaceKineticIntegrand.

Types

AutoBZ.Applications.WannierIntegrand — Type

WannierIntegrand(f, s::AbstractFourierSeries, p)

A type generically representing an integrand f whose entire dependence on the variables of integration is in a Fourier series s, and which may also accept some input parameters p, which are preferrably contained in a tuple. The caller must be aware that their function, f, will be called at many evaluation points, x, in the following way: f(s(x), p...). Therefore the caller is expected to know the type of s(x) (hint: eltype(s)) and the layout of the parameters in the tuple p.

Methods

Missing docstring.

Missing docstring for AutoBZ.Applications.pre_eval_contract(::AutoBZ.Applications.WannierIntegrand,::Any,::Any). Check Documenter's build log for details.

Pre-defined

Types

AutoBZ.Applications.DOSIntegrand — Type

DOSIntegrand(H, ω, Σ, μ)
DOSIntegrand(H, ω, Σ)
DOSIntegrand(H, M)

A type whose integral gives the density of states.

\[D(ω) = -{\pi}^{-1} \int_{\text{BZ}} dk \operatorname{Tr}[\Im[{((\omega + \mu) I - H(k) - \Sigma(\omega))}^{-1}]]\]

This type works with both adaptive and equispace integration routines.

AutoBZ.Applications.TransportIntegrand — Type

TransportIntegrand(HV, Σ, ω₁, ω₂, μ)
TransportIntegrand(HV, Mω₁, Mω₂)

A type whose integral over the BZ gives the transport distribution.

\[\Gamma_{\alpha\beta}(\omega_1, \omega_2) = \int_{\text{BZ}} dk \operatorname{Tr}[\nu_\alpha(k) A(k,\omega_1) \nu_\beta(k) A(k, \omega_2)]\]

This type works with both adaptive and equispace integration routines. Based on TRIQS

AutoBZ.Applications.KineticIntegrand — Type

KineticIntegrand(HV, Σ, β, μ, n, [Ω=0])

A function whose integral over the BZ and the frequency axis gives the kinetic coefficient. Mathematically, this computes

\[A_{n,\alpha\beta}(\Omega) = \int_{-\infty}^{\infty} d \omega (\beta\omega)^{n} \frac{f(\omega) - f(\omega+\Omega)}{\Omega} \Gamma_{\alpha\beta}(\omega, \omega+\Omega)\]

where $f(\omega) = (e^{\beta\omega}+1)^{-1}$ is the Fermi distriubtion. Use this type only for adaptive integration and order the limits so that the integral over the Brillouin zone is the outer integral and the frequency integral is the inner integral. Based on TRIQS

AutoBZ.Applications.EquispaceKineticIntegrand — Type

EquispaceKineticIntegrand(A::KineticIntegrand, l, npt, pre::Vector{Tuple{eltype(A.HV),Int}})
EquispaceKineticIntegrand(A, l, npt)

This type represents an KineticIntegrand, A integrated adaptively in frequency and with equispace integration over the Brillouin zone with a fixed number of grid points npt. The argument l should be an IntegrationLimits for just the Brillouin zone. This type should be called by an adaptive integration routine whose limits of integration are only the frequency variable.

AutoBZ.Applications.AutoEquispaceKineticIntegrand — Type

AutoEquispaceKineticIntegrand(A, l, atol, rtol, npt1, pre1, npt2, pre2)
AutoEquispaceKineticIntegrand(A, l, atol, rtol; npt1=0, pre1=Tuple{eltype(A.HV),Int}[], npt2=0,pre2=Tuple{eltype(σ.HV),Int}[])

This type represents an KineticIntegrand, A integrated adaptively in frequency and with equispace integration over the Brillouin zone with a number of grid points necessary to meet the maximum of the tolerances given by atol and rtol. The argument l should be an IntegrationLimits for just the Brillouin zone. This type should be called by an adaptive integration routine whose limits of integration are only the frequency variable.

The keyword arguments, which are just passed to automatic_equispace_integration, are:

pre1: a Vector containing tuples of the evaluated Hamiltonian + band velocities and integration weights

npt1: an integer that should be equivalent to length(pre1)

pre2: a Vector containing tuples of the evaluated Hamiltonian + band velocities and integration weights on a more refined grid than pre1

npt2: an integer that should be equivalent to length(pre)

Functions

Missing docstring.

Missing docstring for AutoBZ.Applications.dos_integrand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.spectral_function. Check Documenter's build log for details.

AutoBZ.Applications.band_velocities — Function

band_velocities(::Type{Val{kind}}, H, vs...) where kind

Transform the band velocities according to the following values of kind

:full: return the full band velocity (Wannier gauge)

:intra: return only the diagonal of the band velocity (Hamiltonian gauge)

:inter: return only the off-diagonal terms of the band velocity (Hamiltonian gauge)

Missing docstring.

Missing docstring for AutoBZ.Applications.transport_integrand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.kinetic_integrand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.fermi. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.fermi′. Check Documenter's build log for details.

AutoBZ.Applications.fermi_window — Function

Evaluates a unitless window function determined by the Fermi distribution

Missing docstring.

Missing docstring for AutoBZ.Applications.cosh_ratio. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.EXP_P1_SMALL_X. Check Documenter's build log for details.