Integrands

Functions

Missing docstring.

Missing docstring for AutoBZ.Applications.greens_function. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.spectral_function. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.dos_integrand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.gamma_integrand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.oc_integrand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.fermi. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.fermi′. Check Documenter's build log for details.

AutoBZ.Applications.fermi_window — Function

Evaluates a unitless window function determined by the Fermi distribution

Missing docstring.

Missing docstring for AutoBZ.Applications.cosh_ratio. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.EXP_P1_SMALL_X. Check Documenter's build log for details.

Types

AutoBZ.Applications.GreensFunction — Type

GreensFunction(H,ω,Σ,μ)
GreensFunction(H,ω,Σ)
GreensFunction(H,M)

A struct that calculates the lattice Green's function from a Hamiltonian.

\[G(k;H,\omega,\Sigma,\mu) = {((\omega + \mu) I - H(k) - \Sigma(\omega))}^{-1}\]

AutoBZ.Applications.SpectralFunction — Type

SpectralFunction(::GreensFunction)

A struct that calculates the imaginary part of the Green's function.

\[A(k;H,ω,η,μ) = {\pi}^{-1} \Im[G(k;H,ω,η,μ)]\]

AutoBZ.Applications.DOSIntegrand — Type

DOSIntegrand(::SpectralFunction)

A struct whose integral gives the density of states.

\[D(k;H,ω,η,μ) = \operatorname{Tr}[A(k;H,ω,η,μ)]\]

AutoBZ.Applications.GammaIntegrand — Type

GammaIntegrand(H, Σ, ω, Ω, μ)
GammaIntegrand(H, ν₁, ν₂, ν₃, Mω, MΩ)

A function whose integral over the BZ gives the transport distribution.

\[\Gamma_{\alpha\beta}(k) = \operatorname{Tr}[\nu^\alpha(k) A(k,\omega) \nu^\beta(k) A(k, \omega+\Omega)]\]

AutoBZ.Applications.OCIntegrand — Type

OCIntegrand(H, ν₁, ν₂, ν₃, Ω, β, η, μ)

A function whose integral over the BZ and the frequency axis gives the optical conductivity

Missing docstring.

Missing docstring for AutoBZ.Applications.EquispaceOCIntegrand. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Applications.AutoEquispaceOCIntegrand. Check Documenter's build log for details.