Interfaces

Wannier90

AutoBZ.Applications.parse_hamiltonian — Function

parse_hamiltonian(filename)

Parse an ab-initio Hamiltonian output from Wannier90 into filename, extracting the fields (date_time, num_wann, nrpts, degen, irvec, C)

AutoBZ.Applications.load_hamiltonian — Function

load_hamiltonian(filename; period=1.0)

Load an ab-initio Hamiltonian output from Wannier90 into filename as an evaluatable FourierSeries whose periodicity can be set by the keyword argument period which defaults to setting the period along each dimension to 1.0. To define different periods for different dimensions, pass an SVector as the period.

Python

Julia code, including AutoBZ.jl, can be called from Python using the package PyJulia

Setup

TL;DR

$ julia -e 'import Pkg; Pkg.add("PyCall")'
$ python3 -m pip install julia

If you want to, you can install PyJulia in a Python venv, but on the Julia side PyCall.jl must be installed in the default environment

Demo

The Python snippet below shows how from the demos folder of the AutoBZ repository you can run one of the demo scripts:

from julia.api import Julia
jl = Julia(compiled_modules=False)

# julia environment setup in working directory 'demos'
jl.eval("""
import Pkg
Pkg.activate(".")
Pkg.instantiate()
""")

# capture output of script
out = jl.eval('include("DOS_test.jl")')

The first two lines are adapted for loading PyJulia on Debian systems.