Jobs

The module AutoBZ.Jobs has batch functions that setup calculations for WannierIntegrands, DOSIntegrands, and KineticIntegrands.

Parallelized scripts

All of the following scripts have parallelization over parameter points (typically frequencies) enabled by default, except where noted. They also always use $k$ point parallelization of equispace integrals.

Chemical potential

This library does not handle the chemical potential or zero point energy, but recommends that the user offset their Hamiltonian and self-energy data by the chemical potential so that the Fermi level is at zero energy. Doing so, i.e. with AutoBZ.shift!, will ensure a consistent convention and correctness of the kinetic coefficient calculations, which truncate an infinite integral assuming that the Fermi level is at zero energy.

Recommended

The scripts below are recommended because they return a result from adaptive integration, but when rtol>0 they also do an automatic equispace integral as a precomputation to give appropriate absolute tolerances to the adaptive integrator, which performs better with absolute tolerances.

AutoBZ.Jobs.run_wannier — Function

run_wannier(integrand, fs, ps, lims, rtol, atol, [nthreads=Threads.nthreads()]; ertol=1.0, eatol=0.0)

Returns a NamedTuple with names I, E, t, p containing the integrals, errors, and timings for obtained for a Wannier-interpolated integral from caller-supplied Fourier series fs and evaluator integrand done at parameters ps, rtol, atol. The integral is evaluated adaptively. However if rtol is nonzero, it is first estimated by an equispace method using wide tolerances eatol and ertol, and then uses a narrow absolute tolerance set by max(atol,rtol*norm(int)) to evaluate and return the adaptive integral.

AutoBZ.Jobs.run_dos — Function

run_dos(HV, Σ::AbstractSelfEnergy, ωs, BZ_lims, rtol, atol, [nthreads=Threads.nthreads()]; ertol=1.0, eatol=0.0)

Returns a NamedTuple with names I, E, t, omega containing the integrals, errors, and timings for a density of states calculation done at frequencies ωs with parameters atol, rtol. This function first estimates the integral, int, with an equispace method using wide tolerances eatol and ertol, and then uses a narrow absolute tolerance set by max(atol,rtol*norm(int)) to evaluate the same integral adaptively. The domain of integration used is BZ_lims.

Since this is intended to compute a cheap equispace integral first, it is recommended to over-ride the default $k$-grid refinement step to something $\eta$-independent with a line like the one below before calling this script

AutoBZ.equispace_npt_update(npt, ::DOSIntegrand, atol, rtol) = npt + 50

AutoBZ.Jobs.run_kinetic — Function

run_kinetic(HV, Σ::AbstractSelfEnergy, β, n, Ωs, BZ_lims, rtol, atol, [nthreads=Threads.nthreads()]; ertol=1.0, eatol=0.0)

Returns a NamedTuple with names I, E, t, Omega containing the integrals, errors, and timings for a kinetic coefficient calculation done at frequencies Ωs with parameters β, n, atol, rtol. This function constructs both an AutoEquispaceKineticIntegrand with wide tolerances eatol and ertol which estimates the integral, int, and then use a narrow absolute tolerance set by max(atol,rtol*norm(int)) to construct a KineticIntegrand for each parameter value and calls iterated_integration on it over the domain of the BZ_lims and a safely truncated frequency integral to get the results. The calculation is parallelized over Ωs on nthreads threads.

Since this is intended to compute a cheap equispace integral first, it is recommended to over-ride the default $k$-grid refinement step to something $\eta$-independent with a line like the one below before calling this script

AutoBZ.equispace_npt_update(npt, ::TransportIntegrand, atol, rtol) = npt + 50

Adaptive

AutoBZ.Jobs.run_wannier_adaptive — Function

run_wannier_adaptive(integrand, fs, ps, lims, rtol, atol, [nthreads=Threads.nthreads()])

Returns a NamedTuple with names I, E, t, p containing the integrals, errors, and timings for a density of states calculation done at frequencies ωs with parameters atol, rtol. This function constructs a WannierIntegrand for each parameter value and calls iterated_integration with limits of integration lims to get the results.

AutoBZ.Jobs.run_dos_adaptive — Function

run_dos_adaptive(H, Σ::AbstractSelfEnergy, ωs, BZ_lims, rtol, atol, [nthreads=Threads.nthreads()])

Returns a NamedTuple with names I, E, t, omega containing the results, errors, and timings for a density of states calculation done at frequencies ωs with parameters atol, rtol. This function constructs a DOSIntegrand for each parameter value and calls iterated_integration on it over the domain of the BZ_lims to get the results.

AutoBZ.Jobs.run_kinetic_adaptive — Function

run_kinetic_adaptive(HV, Σ::AbstractSelfEnergy, β, n, Ωs, BZ_lims, rtol, atol, [nthreads=Threads.nthreads()])

Returns a NamedTuple with names I, E, t, Omega containing the integrals, errors, and timings for a kinetic coefficient calculation done at frequencies Ωs with parameters β, n, atol, rtol. This function constructs a KineticIntegrand for each parameter value and calls iterated_integration on it over the domain of the BZ_lims and a safely truncated frequency integral to get the results. The calculation is parallelized over Ωs on nthreads threads.

Equispace

The caller supplies a fixed number of $k$ points per dimension, npt, and should check that the solution is converged with respect to this parameter. In general, these routines are useful to get fast results with small numbers of $k$ points (~100) at early stages of a project.

AutoBZ.Jobs.run_wannier_equispace — Function

run_wannier_equispace(integrand, fs, ps, lims, npt, [nthreads=Threads.nthreads()])

Returns a NamedTuple with names I, t, p containing the integrals and timings obtained for a Wannier-interpolated integral from caller-supplied Fourier series fs and evaluator integrand done at parameters ps, npt, where npt is the number of $k$ points per dimension. This function constructs a WannierIntegrand for each parameter value and calls equispace_integration with limits of integration lims to get the results. The caller should check that the result is converged with respect to npt.

AutoBZ.Jobs.run_dos_equispace — Function

run_dos_equispace(HV, Σ::AbstractSelfEnergy, ωs, BZ_lims, npt, [nthreads=Threads.nthreads()])

Returns a NamedTuple with names I, t, omega containing the integrals and timings obtained for a density of states calculation done at frequencies ωs with parameters npt, where npt is the number of $k$ points per dimension. The caller should check that the result is converged with respect to npt. The domain of integration used is BZ_lims.

AutoBZ.Jobs.run_kinetic_equispace — Function

run_kinetic_equispace(HV, Σ::AbstractSelfEnergy, β, n, Ωs, BZ_lims, npt, rtol, atol, [nthreads=Threads.nthreads(), pre_eval=pre_eval_contract])

Returns a NamedTuple with names I, E, t, Omega containing the integrals, errors, and timings for a kinetic coefficient calculation done at frequencies Ωs with parameters β, n, atol, rtol. This function constructs an EquispaceKineticIntegrand for each parameter value, and precomputes HV on an equispace $k$ grid with npt points per dimension (which is reused for all parameter values), and calls iterated_integration on it over the domain of the BZ_lims and a safely truncated frequency integral to get the results. The calculation is parallelized over Ωs on nthreads threads.

Automatic equispace

Frequency parallelization is not enabled by default due to the likelihood of running out of memory large $k$ point grids.

AutoBZ.Jobs.run_wannier_auto_equispace — Function

run_wannier_auto_equispace(integrand, fs, ps, lims, rtol, atol, [nthreads=1])

Returns a NamedTuple with names I, E, t, p, npt1, npt2 containing the integrals, errors, timings, and number of kpts per dimension obtained for a Wannier-interpolated integral from caller-supplied Fourier series fs and evaluator integrand done at parameters ps, atol, rtol. This function constructs a WannierIntegrand for each parameter value and calls automatic_equispace_integration with limits of integration lims to get the results.

AutoBZ.Jobs.run_dos_auto_equispace — Function

run_dos_auto_equispace(HV, Σ::AbstractSelfEnergy, ωs, BZ_lims, rtol, atol, [nthreads=1])

Returns a NamedTuple with names I, E, t, omega, npt1, npt2 containing the integrals, errors, timings, and number of kpts per dimension used for a density of states calculation done at frequencies ωs with parameters atol, rtol. The domain of integration used is BZ_lims.

AutoBZ.Jobs.run_kinetic_auto_equispace — Function

run_kinetic_auto_equispace(HV, Σ::AbstractSelfEnergy, β, n, Ωs, BZ_lims, rtol, atol, [nthreads=1])

Returns a NamedTuple with names I, E, t, Omega, npt1, npt2 containing the integrals, errors, timings, and kpts used for a kinetic coefficient calculation done at frequencies Ωs with parameters β, n, atol, rtol. This function constructs an AutoEquispaceKineticIntegrand for each parameter value, reusing $k$-grids of HV values from previous calculations, and calls iterated_integration on it over the domain of the BZ_lims and a safely truncated frequency integral to get the results. The calculation is parallelized over Ωs on nthreads threads. The default is set to 1 thread for frequency parallelization, although k-point parallelization is still enabled, to avoid duplicating calculations of the Hamiltonian and band velocities on the k-mesh.

I/O

AutoBZ.Jobs.read_h5_to_nt — Function

read_h5_to_nt(filename)

Loads the h5 archive from filename and reads its datasets into a NamedTuple and its groups into NamedTuples recursively.

AutoBZ.Jobs.write_nt_to_h5 — Function

write_nt_to_h5(nt::NamedTuple, filename)

Takes a NamedTuple and writes its values, which must be arrays, into an h5 archive at filename with dataset names corresponding to the tuple names. If a value is a NamedTuple, its datasets are written to h5 groups recursively.

Internal

AutoBZ.Jobs.get_safe_freq_limits — Function

get_safe_freq_limits(Ωs, β, lb, ub)

Given a collection of frequencies, Ωs, returns a Vector{CubicLimits{1}} with truncated limits of integration for the frequency integral at each (Ω, β) point that are determined by the fermi_window_limits routine set to the default tolerances for the decay of the Fermi window function. The arguments lb and ub are lower and upper limits on the frequency to which the default result gets truncated if the default result would recommend a wider interval. If there is any truncation, a warning is emitted to the user, but the program will continue with the truncated limits.

AutoBZ.Jobs.batch_smooth_param — Function

batch_smooth_param(xs, nthreads)

If the cost of a calculation smoothly varies with the parameters xs, then batch xs into nthreads groups where the ith element of group j is xs[j+(i-1)*nthreads]