Jobs

The module AutoBZ.Jobs has batch functions that setup optical conductivity calculations.

Scripts

Missing docstring.

Missing docstring for AutoBZ.Jobs.OCscript. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Jobs.OCscript_equispace. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Jobs.OCscript_auto. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Jobs.OCscript_auto_equispace. Check Documenter's build log for details.

Parallelized scripts

Missing docstring.

Missing docstring for AutoBZ.Jobs.OCscript_parallel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Jobs.OCscript_equispace_parallel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Jobs.OCscript_auto_parallel. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AutoBZ.Jobs.OCscript_auto_equispace_parallel. Check Documenter's build log for details.

I/O

Note: currently Julia's HDF5 module handles arrays of complex values inconsistently with Python's HDF5 module. This could be fixed by simply using PyCall to use Python's h5py library instead.

AutoBZ.Jobs.read_h5_to_nt — Function

read_h5_to_nt(filename)

Loads the h5 archive from filename and reads its datasets into a NamedTuple

AutoBZ.Jobs.write_nt_to_h5 — Function

write_nt_to_h5(nt::NamedTuple, filename)

Takes a NamedTuple and writes its values, which must be arrays, into an h5 archive at filename with dataset names corresponding to the tuple names.

AutoBZ.Jobs.import_self_energy — Function

import_self_energy(filename)

Reads the groups omega and sigma in the h5 archive in filename and tries save it to a NamedTuple with names ω and Σ. The array in sigma should be of size (length(omega), 2), where the two columns are the real and imaginary parts of Σ.

Internal

Missing docstring.

Missing docstring for AutoBZ.Jobs.BandEnergyBerryVelocities. Check Documenter's build log for details.

AutoBZ.Jobs.get_safe_freq_limits — Function

get_safe_freq_limits(Ωs, β, lb, ub)

Given a collection of frequencies, Ωs, returns a Vector{CubicLimits{1}} with truncated limits of integration for the frequency integral at each (Ω, β) point that are determined by the fermi_window_limits routine set to the default tolerances for the decay of the Fermi window function. The arguments lb and ub are lower and upper limits on the frequency to which the default result gets truncated if the default result would recommend a wider interval. If there is any truncation, a warning is emitted to the user, but the program will continue with the truncated limits.

AutoBZ.Jobs.batch_smooth_param — Function

batch_smooth_param(xs, nthreads)

If the cost of a calculation smoothly varies with the parameters xs, then batch xs into nthreads groups where the ith element of group j is xs[j+(i-1)*nthreads]