Jobs

The module AutoBZ.Jobs has batch functions that setup optical conductivity calculations.

Scripts

AutoBZ.Jobs.OCscript — Function

OCscript(HV::BandEnergyBerryVelocities, Σ::AbstractSelfEnergy, β, Ωs, μ, atol, rtol)

Returns a NamedTuple with names OC, err, t, Omega containing the results, errors, and timings for an optical conductivity calculation done at frequencies Ωs with parameters β, μ, atol, rtol. This function constructs an OCIntegrand for each parameter value and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results.

AutoBZ.Jobs.OCscript_equispace — Function

OCscript_equispace(HV::BandEnergyBerryVelocities, Σ::AbstractSelfEnergy, β, Ωs, μ, npt, atol, rtol; pre_eval=pre_eval_contract)

Returns a NamedTuple with names OC, err, t, Omega containing the results, errors, and timings for an optical conductivity calculation done at frequencies Ωs with parameters β, μ, atol, rtol. This function constructs an EquispaceOCIntegrand for each parameter value, and precomputes HV on an equispace $k$ grid with npt points per dimension (which is reused for all parameter values), and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results.

AutoBZ.Jobs.OCscript_auto — Function

OCscript_auto(HV::BandEnergyBerryVelocities, Σ::AbstractSelfEnergy, β, Ωs, μ, atol, rtol; ertol=1.0, eatol=0.0)

Returns a NamedTuple with names OC, err, t, Omega containing the results, errors, and timings for an optical conductivity calculation done at frequencies Ωs with parameters β, μ, atol, rtol. This function constructs both an AutoEquispaceOCIntegrand with wide tolerances eatol and ertol which estimates the integral, int, and then uses a narrow absolute tolerance set by max(atol,rtol*norm(int)) to construct an OCIntegrand for each parameter value and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results.

Since this is intended to compute a cheap equispace integral first, it is recommended to over-ride the default $k$-grid refinement step to something $\eta$-independent with a line like the one below before calling this script

AutoBZ.equispace_npt_update(npt, ::GammaIntegrand, atol, rtol) = npt + 50

AutoBZ.Jobs.OCscript_auto_equispace — Function

OCscript_auto_equispace(HV::BandEnergyBerryVelocities, Σ::AbstractSelfEnergy, β, Ωs, μ, atol, rtol)

Returns a NamedTuple with names OC, err, t, Omega containing the results, errors, and timings for an optical conductivity calculation done at frequencies Ωs with parameters β, μ, atol, rtol. This function constructs an AutoEquispaceOCIntegrand for each parameter value, reusing $k$-grids of HV values from previous calculations, and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results.

Parallelized scripts

AutoBZ.Jobs.OCscript_parallel — Function

OCscript_parallel(filename, HV::BandEnergyBerryVelocities, Σ::AbstractSelfEnergy, β, Ωs, μ, atol, rtol; nthreads=Threads.nthreads())

Writes an h5 archive to filename with groups OC, err, t, Omega containing the results, errors, and timings for an optical conductivity calculation done at frequencies Ωs with parameters β, μ, atol, rtol. This function constructs an OCIntegrand for each parameter value and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results. The calculation is parallelized over Ωs on nthreads threads.

AutoBZ.Jobs.OCscript_equispace_parallel — Function

OCscript_equispace_parallel(filename, HV::BandEnergyBerryVelocities, Σ::AbstractSelfEnergy, β, Ωs, μ, npt, atol, rtol; pre_eval=pre_eval_contract, nthreads=Threads.nthreads())

Writes an h5 archive to filename with groups OC, err, t, Omega containing the results, errors, and timings for an optical conductivity calculation done at frequencies Ωs with parameters β, μ, atol, rtol. This function constructs an EquispaceOCIntegrand for each parameter value, and precomputes HV on an equispace $k$ grid with npt points per dimension (which is reused for all parameter values), and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results. The calculation is parallelized over Ωs on nthreads threads.

AutoBZ.Jobs.OCscript_auto_parallel — Function

OCscript_auto_parallel(HV::BandEnergyBerryVelocities, Σ::AbstractSelfEnergy, β, Ωs, μ, atol, rtol; ertol=1.0, eatol=0.0, nthreads=Threads.nthreads())

Returns a NamedTuple with names OC, err, t, Omega containing the results, errors, and timings for an optical conductivity calculation done at frequencies Ωs with parameters β, μ, atol, rtol. This function constructs both an AutoEquispaceOCIntegrand with wide tolerances eatol and ertol which estimates the integral, int, and then use a narrow absolute tolerance set by max(atol,rtol*norm(int)) to construct an OCIntegrand for each parameter value and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results. The calculation is parallelized over Ωs on nthreads threads.

Since this is intended to compute a cheap equispace integral first, it is recommended to over-ride the default $k$-grid refinement step to something $\eta$-independent with a line like the one below before calling this script

AutoBZ.equispace_npt_update(npt, ::GammaIntegrand, atol, rtol) = npt + 50

AutoBZ.Jobs.OCscript_auto_equispace_parallel — Function

OCscript_auto_equispace_parallel(filename, HV::BandEnergyBerryVelocities, Σ::AbstractSelfEnergy, β, Ωs, μ, atol, rtol; nthreads=Threads.nthreads())

Writes an h5 archive to filename with groups OC, err, t, Omega containing the results, errors, and timings for an optical conductivity calculation done at frequencies Ωs with parameters β, μ, atol, rtol. This function constructs an AutoEquispaceOCIntegrand for each parameter value, reusing $k$-grids of HV values from previous calculations, and calls iterated_integration on it over the domain of the IBZ and a safely truncated frequency integral to get the results. The calculation is parallelized over Ωs on nthreads threads.

I/O

Note: currently Julia's HDF5 module handles arrays of complex values inconsistently with Python's HDF5 module. This could be fixed by simply using PyCall to use Python's h5py library instead.

AutoBZ.Jobs.read_h5_to_nt — Function

read_h5_to_nt(filename)

Loads the h5 archive from filename and reads its datasets into a NamedTuple

AutoBZ.Jobs.write_nt_to_h5 — Function

write_nt_to_h5(nt::NamedTuple, filename)

Takes a NamedTuple and writes its values, which must be arrays, into an h5 archive at filename with dataset names corresponding to the tuple names.

AutoBZ.Jobs.import_self_energy — Function

import_self_energy(filename)

Reads the groups omega and sigma in the h5 archive in filename and tries save it to a NamedTuple with names ω and Σ. The array in sigma should be of size (length(omega), 2), where the two columns are the real and imaginary parts of Σ.

Internal

AutoBZ.Jobs.BandEnergyBerryVelocities — Type

BandEnergyBerryVelocities

Union type of BandEnergyBerryVelocity, BandEnergyBerryVelocity3D, BandEnergyVelocity, and BandEnergyVelocity3D.

AutoBZ.Jobs.get_safe_freq_limits — Function

get_safe_freq_limits(Ωs, β, lb, ub)

Given a collection of frequencies, Ωs, returns a Vector{CubicLimits{1}} with truncated limits of integration for the frequency integral at each (Ω, β) point that are determined by the fermi_window_limits routine set to the default tolerances for the decay of the Fermi window function. The arguments lb and ub are lower and upper limits on the frequency to which the default result gets truncated if the default result would recommend a wider interval. If there is any truncation, a warning is emitted to the user, but the program will continue with the truncated limits.

AutoBZ.Jobs.batch_smooth_param — Function

batch_smooth_param(xs, nthreads)

If the cost of a calculation smoothly varies with the parameters xs, then batch xs into nthreads groups where the ith element of group j is xs[j+(i-1)*nthreads]