Adaptive integration

Routines

AutoBZ.tree_integration — Function

tree_integration(f, a, b)
tree_integration(f, ::CubicLimits)

Calls HCubature to perform multi-dimensional integration of f over a cube.

AutoBZ.iterated_integration — Function

iterated_integration(f, ::IntegrationLimits)

Calls HCubature to perform iterated 1D integration of f over a domain parametrized by IntegrationLimits. Accepts a callback function whose arguments are f and the evaluation point, x, as a keyword argument. The callback can return a modified integrand to the next inner integral, but the default is thunk which delays the computation to the innermost integral.

Customization

AutoBZ.thunk — Function

thunk(f, x)

Delay the computation of f(x). Needed to normally evaluate an integrand in nested integrals as employed by callbacks.

AutoBZ.ThunkIntegrand — Type

ThunkIntegrand(f, x)

Store f and x to evaluate f(x) at a later time. Can be employed in iterated integration

AutoBZ.default_error_callback — Function

Choose a new set of error tolerances for the next inner integral.