Fourier series

Wannier-interpolated Hamiltonians are represented by Fourier series with a modest number of Fourier coefficients. The goal of this page of documentation is to describe the features, interface, and conventions of Fourier series evaluation as implemented by this library.

Conventions

Fourier series represent functions as linear combinations of sinusoids whose frequencies are integer multiples of a fundamental mode. In the band theory of solids, the fundamental frequencies, or normal modes, for a Hamiltonian correspond to linear combinations of real-space lattice vectors that generate a Bravais lattice. The sections below define conventions for each of these linear combinations.

Lattice vectors

It is conventional to construct a reciprocal lattice $\{\bm{b}_j\}$ from a Bravais lattice $\{\bm{a}_i\}$ such that

\[\bm{b}_j \cdot \bm{a}_i = 2\pi\delta_{ij}\]

Then we write the momentum space variable in the reciprocal lattice vector basis and the position space variable in the real lattice vector basis

\[\bm{k} = \sum_{j=1}^{d} k_j \bm{b}_j \qquad \bm{R} = \sum_{i=1}^{d} R_i \bm{a}_i\]

Without loss of generality, the $k_j$ can be taken in the domain $[0,1]$ and the $R_i$ are integers.

Additionally, any coordinate transformations of $\bm{k}$ from the Cartesian basis to the reciprocal lattice basis only modify Brillouin zone integrals by a multiplicative factor of the absolute value of the determinant of the basis transformation. To find a non-trivial example of representing a Hamiltonian in the reciprocal lattice basis, see the Graphene example with ManyOffsetsFourierSeries.

Series coefficients

Wannier-interpolated Hamiltonians are small matrices obtained by projecting the ab-initio Hamiltonian onto a low energy subspace, a process called downfolding. These Hamiltonians can be expressed by a Fourier series as in the sum below

\[H(\bm{k}) = \sum_{\bm{R}} e^{i\bm{k}\cdot\bm{R}} H_{\bm{R}}\]

where the coefficients $H_{\bm{R}}$ are the matrix-valued Fourier coefficients. Truncating the sum over $\bm{R}$ at a modest number of modes can be done for Wannier Hamiltonians in the maximally-localized orbital basis, for which $H(\bm{k})$ is a smooth and periodic function and thus the truncation error of its Fourier series converges super-algebraically with respect to the number of modes.

Hamiltonian recipe

In model systems, a Bloch Hamiltonian can often be written down analytically. The recipe to write it as a Fourier series has two-steps

Identify the real and reciprocal Bravais lattices, $\{\bm{a}_i\}$ and $\{\bm{b}_j\}$, and rewrite all of the phase dependences of the Hamiltonian as $\bm{k}\cdot\bm{R}$ with each vector in its corresponding basis, as explained above.

Factor the Hamiltonian into a linear combination of normal modes indexed by the distinct $\bm{R}$ vectors. If the Hamiltonian is matrix-valued, this can be done one matrix element at a time.

Interface

AutoBZ.AbstractFourierSeries — Type

AbstractFourierSeries{N}

A supertype for Fourier series that are periodic maps $\R^N \to V$ where $V$ is any vector space. Typically these can be represented by N-dimensional arrays whose elements belong to the vector space. See the manual section on the AbstractFourierSeries interface.

AutoBZ.period — Function

period(f::AbstractFourierSeries{N}) where {N}

Return a NTuple{N} whose m-th element corresponds to the period of f along its m-th input dimension. Typically, these values set the units of length for the problem.

AutoBZ.contract — Function

contract(f::AbstractFourierSeries{N}, x::Number, [dim=N]) where {N}

Return another Fourier series of dimension N-1 by summing over dimension dim of f with the phase factors evaluated at x. If N=1, this function should return an AbstractFourierSeries{0} that stores the evaluated Fourier series, but has no more input dimensions to contract.

The default of dim=N is motivated by preserving memory locality in Julia's column-major array format.

contract(f::AbstractFourierSeries{N}, x::SVector{M}) where {N,M}

Contract the outermost indices M of f in order of last(x) to first(x). If M>N, the default behavior is just to try and contract M indices, which will likely lead to an error.

AutoBZ.value — Function

value(::AbstractFourierSeries{0})

Return the evaluated Fourier series whose indices have all been contracted. Typically, this value has the same units as the Fourier series coefficients.

Additonally, concrete subtypes of AbstractFourierSeries must have an element type, which they can do by extending Base.eltype with a method. For example, if a type MyFourierSeries <: AbstractFourierSeries always returns ComplexF64 outputs, then the correct eltype method to define would be:

Base.eltype(::Type{MyFourierSeries}) = ComplexF64

The type returned should correspond to the vector space $V$ of the output space of the Fourier series, i.e. the output of value should be of this type. For good performance, the eltype should be a concrete type and should be inferrable.

With the above implemented, several methods which define functors for AbstractFourierSeries allow the user (and integration routines) to evaluate the type like a function with the f(x) syntax.

Types

The concrete types listed below all implement the AbstractFourierSeries interface and should cover most use cases.

AutoBZ.FourierSeries — Type

FourierSeries(coeffs, period::SVector{N,Float64}) where {N}

Construct a Fourier series whose coefficients are given by the coefficient array array coeffs whose eltype should support addition and scalar multiplication, and whose periodicity on the ith axis is given by period[i]. This type represents the Fourier series

\[f(\vec{x}) = \sum_{\vec{n} \in \mathcal I} C_{\vec{n}} \exp(i2\pi\vec{k}_{\vec{n}}\cdot\overrightarrow{x})\]

where $i = \sqrt{-1}$ is the imaginary unit, $C$ is the array coeffs, $\mathcal I$ is CartesianIndices(C), $\vec{n}$ is a CartesianIndex and $\vec{k}_{\vec{n}}$ is equal to $n_j/p_j$ in the $j$th position with $p_j$ the $j$th element of period. Because of the choice to use Cartesian indices to set the phase factors, typically the indices of coeffs should be specified by using an OffsetArray.

FourierSeries(coeffs::AbstractArray{T,N}, period::Real) where {T,N}

If period is a Real, this constructor will infer the number of input dimensions of the Fourier series from the array dimensionality of the coefficients, and period will become the period of all of the dimensions.

AutoBZ.FourierSeriesDerivative — Type

FourierSeriesDerivative(f::FourierSeries{N}, a::SVector{N}) where {N}

Represent the differential of Fourier series f by a multi-index a of derivatives, e.g. [1,2,...], whose ith entry represents the order of differentiation on the ith input dimension of f. Mathematically, this means

\[\left( \prod_{j=1}^N \partial_{x_j}^{a_j} \right) f(\vec{x}) = \sum_{\vec{n} \in \mathcal I} \left( \prod_{j=1}^N (i 2\pi k_j)^{a_j} \right) C_{\vec{n}} \exp(i2\pi\vec{k}_{\vec{n}}\cdot\overrightarrow{x})\]

where $\partial_{x_j}^{a_j}$ represents the $a_j$th derivative of $x_j$, $i = \sqrt{-1}$ is the imaginary unit, $C$ is the array coeffs, $\mathcal I$ is CartesianIndices(C), $\vec{n}$ is a CartesianIndex and $\vec{k}_{\vec{n}}$ is equal to $n_j/p_j$ in the $j$th position with $p_j$ the $j$th element of period. Because of the choice to use Cartesian indices to set the phase factors, typically the indices of coeffs should be specified by using an OffsetArray. Also, note that termwise differentiation of the Fourier series results in additional factors of $i2\pi$ which should be anticipated for the use case. Also, note that this type can be used to represent fractional differentiation or integration by suitably choosing the $a_j$s.

This is a 'lazy' representation of the derivative because instead of differentiating by computing all of the Fourier coefficients of the derivative upon constructing the object, the evaluator waits until it contracts the differentiated dimension to evaluate the new coefficients.

AutoBZ.OffsetFourierSeries — Type

OffsetFourierSeries(f::AbstractFourierSeries{N}, q::SVector{N,Float64}) where {N}

Represent a Fourier series whose argument is offset by the vector $\vec{q}$ and evaluates it as $f(\vec{x}-\vec{q})$.

AutoBZ.ManyFourierSeries — Type

ManyFourierSeries(fs::AbstractFourierSeries{N}...) where {N}

Represents a tuple of Fourier series of the same dimension and periodicity and contracts them all simultaneously.

AutoBZ.ManyOffsetsFourierSeries — Type

ManyOffsetsFourierSeries(f, qs..., [origin=true])

Represent a Fourier series evaluated at many different points, and contract them all simultaneously, returning them in the order the qs were passed, i.e. (f(x-qs[1]), f(x-qs[2]), ...) The origin keyword decides whether or not to evaluate $f$ without an offset, and if origin is true, the value of $f$ evaluated without an offset will be returned in the first position of the output.

AutoBZ.BandEnergyVelocity — Type

BandEnergyVelocity{kind}(H::FourierSeries{N}, V::ManyFourierSeries{N}) where {kind,N}
BandEnergyVelocity(H::FourierSeries{N}; kind=:orbital) where N

The bottom constructor takes a Fourier series representing the Hamiltonian and also evaluates the band velocities so that the return value after all the dimensions are contracted is a tuple containing (H, v₁, v₂, ..., vₙ). The band velocities are defined by dipole operators $\nu_{\alpha} = \frac{1}{\hbar} \partial_{k_{\alpha}} H$ where $k_{\alpha}$ is one of three input dimensions of $H$ and $\hbar=1$. Effectively, this type evaluates H and its gradient. Note that differentiation by $k$ changes the units to have an additional dimension of length and a factor of $2\pi$, so if $H$ has dimensions of energy, $\nu$ has dimensions of energy times length. The caller is responsible for transforming the units of the velocity (i.e. $\hbar$) if they want other units, which can usually be done as a post-processing step.

AutoBZ.BandEnergyBerryVelocity — Type

BandEnergyBerryVelocity{kind}(H::BandEnergyVelocity{Val{:orbital}(),N}, A::ManyFourierSeries{N}) where {kind,N}
BandEnergyBerryVelocity(H::BandEnergyVelocity{kind,N}, A::ManyFourierSeries{N,<:NTuple{N}}) where {kind,N}
BandEnergyBerryVelocity(H::FourierSeries, A; kind=:orbital)

This constructor takes a FourierSeries, H, representing the Hamiltonian and also a ManyFourierSeries, A, representing the gradient of the Berry connection, and evaluates modified band velocities so that the return value after all the dimensions are contracted is a tuple containing (H, ṽ₁, ṽ₂, ..., ṽₙ). The modified band velocities are defined by

\[\tilde{\nu}_{\alpha} = \frac{1}{\hbar} \partial_{k_{\alpha}} H - \frac{i}{\hbar} [H,(A_{\alpha} + A_{\alpha}^{\dagger})/2]\]

where $k_{\alpha}$ is one of three input dimensions of $H$ and $\hbar=1$. Effectively, this type evaluates the Hamiltonian and its gradient modified by a commutator of the Hamiltonian with the gradient of the Berry connection. Note that differentiation by $k$ changes the units to have an additional dimension of length and a factor of $2\pi$, so if $H$ has dimensions of energy, $\nu$ has dimensions of energy times length. The caller is responsible for transforming the units of the velocity (i.e. $\hbar$) if they want other units, which can usually be done as a post-processing step.

Methods

AutoBZ.contract — Method

contract(f::FourierSeries{N}, x::Number, [dim=N]) where N

Contract index dim of the coefficients of f at the spatial point x. The default dim is the outermost dimension to preserve memory locality.

There is a special optimization when dim=N, size(f.coeffs,dim)=2M+1 for an integer M, and axes(f.coeffs,dim)=-M:M that evaluates the Fourier series by recurrence. In this case, f.coeffs must be an OffsetArray with odd axis lengths and indices that are odd about the center of the array.

contract(f::FourierSeriesDerivative{N}, x::Number, [dim=N]) where {N}

Contract index dim of the coefficients of f at the spatial point x. The default dim is the outermost dimension to preserve memory locality.

There is a special optimization when dim=N, size(f.f.coeffs,dim)=2M+1 for an integer M, and axes(f.f.coeffs,dim)=-M:M that evaluates the Fourier series by recurrence. In this case, f.f.coeffs must be an OffsetArray with odd axis lengths and indices that are odd about the center of the array.

Optimized 3D evaluators

For the use-case of Wannier90 calculations, the following Fourier series evaluators are optimized to improve performance by reducing allocations.

Interface

AutoBZ.AbstractFourierSeries3D — Type

AbstractFourierSeries3D <: AbstractFourierSeries{3}

An abstract subtype of AbstractFourierSeries{3} representing in-place Fourier series evaluators

AutoBZ.contract! — Function

contract!(f::AbstractFourierSeries3D, x::Number, dim::Int)

An in-place version of contract.

AutoBZ.shift! — Function

shift!(f::FourierSeries3D, λ::Number)

Modifies and returns f such that it returns f - λ*I.

Types

AutoBZ.FourierSeries3D — Type

FourierSeries3D(coeffs::Array{T,3}, [period=(1.0, 1.0, 1.0)])

This type is an AbstractFourierSeries{3} designed for in-place evaluation of FourierSeries, and unlike FourierSeries is specialized for 3D Fourier series and does not allocate a new array every time contract is called on it. This type stores the intermediate arrays used in a calculation and assumes that the size of coeffs on each axis is odd because it treats the zeroth harmonic as the center of the array (i.e. (size(coeffs) .÷ 2) .+ 1).

AutoBZ.BandEnergyVelocity3D — Type

BandEnergyVelocity3D(coeffs, [period=(1.0,1.0,1.0), kind=:full])
BandEnergyVelocity3D(H::FourierSeries3D, [kind=:full])

The in-place equivalent of BandEnergyVelocity for 3D series evaluation.

AutoBZ.BandEnergyBerryVelocity3D — Type

BandEnergyBerryVelocity3D(H::FourierSeries3D{TH,0,0,0}, Ax::FourierSeries3D{TA,0,0,0}, Ay::FourierSeries3D{TA,0,0,0}, Az::FourierSeries3D{TA,0,0,0}, [kind=:full]) where {TH,TA}

The in-place equivalent of BandEnergyBerryVelocity for 3D series evaluation.

Methods

Missing docstring.

Missing docstring for AutoBZ.contract!(::AutoBZ.AbstractFourierSeries3D). Check Documenter's build log for details.

AutoBZ.fourier_kernel! — Function

fourier_kernel!(r::Array{T,N-1}, C::Array{T,N}, x, ξ, [::Val{a}=Val{0}()]) where {T,N,a}

Contract the outermost index of array C and write it to the array r. Assumes the size of the outermost dimension of C is 2m+1 and sums the coefficients

\[r_{i_{1},\dots,i_{N-1}} = \sum_{i_{N}=-m}^{m} C_{i_{1},\dots,i_{N-1},i_{N}+m+1} (i2\pi\xi i_{N})^{a} \exp(i2\pi\xi x i_{N})\]

Hence this represents evaluation of a Fourier series with m modes. The parameter a represents the order of derivative of the Fourier series.