Internal

The following symbols used by AutoBZ may change in future versions

AutoBZ.SSymmetricCompact — Type

SSymmetricCompact{N, T, L} <: StaticMatrix{N, N, T}

A StaticArray subtype that represents a Symmetric matrix. Unlike LinearAlgebra.Symmetric, SSymmetricCompact stores only the lower triangle of the matrix (as an SVector). The lower triangle is stored in column-major order. For example, for an SSymmetricCompact{3}, the indices of the stored elements can be visualized as follows:

┌ 1 ⋅ ⋅ ┐
| 2 4 ⋅ |
└ 3 5 6 ┘

Type parameters:

N: matrix dimension;

T: element type for lower triangle;

L: length of the SVector storing the lower triangular elements.

Note that L is always the Nth triangular number.

An SSymmetricCompact may be constructed either:

from an AbstractVector containing the lower triangular elements; or

from a Tuple containing both upper and lower triangular elements in column major order; or

from another StaticMatrix.

For the latter two cases, only the lower triangular elements are used; the upper triangular elements are ignored.

AutoBZ.hinv — Function

hinv(A::SHermitianCompact{3})

Calculate the inverse of a SHermitianCompact matrix using its lower triangle. Note that if the elements on the diagonal are complex, the inverse is not Hermitian.

AutoBZ.dos_integrand — Function

dos_integrand(h, M)

Returns -imag(tr(inv(M-h)))/pi where M = ω*I-Σ(ω). It is unsafe to use this in the inner integral for small eta due to the localized tails of the integrand. The default, safe version also integrates the real part which is less localized, at the expense of a slight slow-down due to complex arithmetic. See TrGlocIntegrand.

AutoBZ.AbstractGauge — Type

AbstractGauge

Abstract supertype of gauges (or bases) for orbital/band indices

AutoBZ.Hamiltonian — Type

Hamiltonian <: AbstractGauge

Singleton type representing the Hamiltonian gauge

AutoBZ.tr_kron — Function

tr_kron(A::T, B::T) where {T<:SVector{AbstractMatrix}}

Returns a matrix whose [i,j]th entry is tr(A[i]*B[j]).

AutoBZ.gloc_integrand — Function

gloc_integrand(h, M)

Returns inv(M-h) where M = ω*I-Σ(ω)

AutoBZ.AbstractGaugeInterp — Type

AbstractGaugeInterp{G,N,T,iip} <: AbstractWannierInterp{N,T,iip}

An abstract subtype of AbstractFourierSeries representing in-place Fourier series evaluators for Wannier-interpolated quantities with a choice of basis, or gauge, G, which is typically Hamiltonian or Wannier. For details, see to_gauge.

AutoBZ.diag_inv — Function

diag_inv(A)

Calculate the diagonal entries of the inverse of A.

AutoBZ.AutoBZ — Module

An applications package implementing iterated-adaptive integration and equispace integration for electronic structure and transport calculations. It excels at integrating both localized and also broadened Brillouin-zone integrands, using the algorithms described by Kaye et al.. This package also provides multi-threaded routines for parallelized calculations. See AutoBZCore if you only need the essential functionality of the library to define custom BZ integrands.

AutoBZ.ConstScalarSelfEnergy — Type

ConstScalarSelfEnergy(v::Number)

Construct a self-energy evaluator which returns $v I$ for any frequency.

AutoBZ.tr_mul — Function

tr_mul(A, B)

Calculate tr(A*B) without storing the intermediate result.

AutoBZ.to_gauge — Function

to_gauge(::AbstractGauge, h) where gauge

Transform the Hamiltonian according to the following values of gauge

Wannier: keeps h, vs in the original, orbital basis

Hamiltonian: diagonalizes h and rotates h into the energy, band basis

AutoBZ.AbstractVelocityInterp — Type

AbstractVelocityInterp{C,B,G,N,T,iip} <:AbstractCoordInterp{B,G,N,T,iip}

An abstract subtype of AbstractCoordInterp also containing information the velocity component, C, which is typically Val(:whole), Val(:inter), or Val(:intra). For details see to_vcomp_gauge. Since the velocity depends on the Hamiltonian, subtypes should also evaluate the Hamiltonian.

AutoBZ.fermi_window_halfwidth — Function

fermi_window_halfwidth(Ω, β, atol)
fermi_window_halfwidth(β, atol)

One can show that βΩfermiwindow(ω, β, Ω) = -tanh(βΩ/2)/(cosh(β(ω+Ω/2))/cosh(βΩ/2)+1) > -tanh(βΩ/2)/(exp(abs(β(ω+Ω/2)))/2cosh(βΩ/2)+1) as well as when Ω==0, β*fermiwindow(ω, β, 0.0) = and these can be inverted to give a good bound on the width of the frequency window for which the Fermi window function is greater than atol. Returns half the width of this window.

AutoBZ.diaggloc_integrand — Function

diaggloc_integrand(h, M)

Returns diag(inv(M-h)) where M = ω*I-Σ(ω)

AutoBZ.Wannier — Type

Wannier <: AbstractGauge

Singleton type representing the Wannier gauge

AutoBZ.DiagonalSelfEnergy — Type

DiagonalSelfEnergy(interpolant, lb, ub)

Construct a self-energy evaluator which for frequencies above lb and below ub returns the vector interpolant at that frequency wrapped by a Diagonal.

AutoBZ.CoordDefault — Function

CoordDefault(::Type{T})::AbstractCoordinate where T

AbstractCoordInterps should define this trait to declare the coordinate basis where they assume their data is in.

AutoBZ.CartesianRep — Type

CartesianRep()

Symmetry representation of objects that transform under the group action in the same way as the lattice and in Cartesian coordinates.

AutoBZ.to_coord — Function

to_coord(B::AbstractCoordinate, D::AbstractCoordinate, A, vs)

If B and D are the same type return vs, however and if they differ return A*vs.

AutoBZ.AbstractWannierInterp — Type

AbstractWannierInterp{N,T,iip} <: AbstractFourierSeries{N,T,iip}

Abstract supertype for all Wannier-interpolated quantities in AutoBZ

AutoBZ.AbstractVelocityComponent — Type

AbstractVelocityComponent

Abstract supertype representing velocity components.

AutoBZ.AbstractCoordInterp — Type

AbstractCoordInterp{B,G,N,T,iip} <:AbstractGaugeInterp{G,N,T,iip}

An abstract subtype of AbstractGaugeInterp also containing information about the coordinate basis B, which is either Lattice or Cartesian. For details see to_coord and CoordDefault.

AutoBZ.shift! — Function

shift!(h::AbstractHamiltonianInterp, λ::Number)

Modifies and returns h such that it returns h - λ*I. Will throw a BoundsError if this operation cannot be done on the existing data.

shift!(::AbstractVelocityInterp, λ)

Offset the zero-point energy in a Hamiltonian system by a constant

AutoBZ.AbstractHamiltonianInterp — Type

AbstractHamiltonianInterp{G,N,T,iip} <: AbstractGaugeInterp{G,N,T,iip}

Abstract type representing Hamiltonians, which are matrix-valued Hermitian Fourier series. They should also have period 1, but produce derivatives with wavenumber 1.

AutoBZ.AbstractCoordinate — Type

AbstractCoordinate

Abstract supertype of bases for coordinate/spatial indices

AutoBZ.Cartesian — Type

Cartesian <: AbstractCoordinate

Singleton type representing Cartesian coordinates.

AutoBZ.tr_inv — Function

tr_inv(A)

Calculate the trace of the inverse of A.

AutoBZ.to_vcomp_gauge — Function

to_vcomp_gauge(::Val{C}, ::Val{G}, h, vs...) where {C,G}

Take the velocity components of vs in any gauge according to the value of C

Whole: return the whole velocity (sum of interband and intraband components)

Intra: return the intraband velocity (diagonal in Hamiltonian gauge)

Inter: return the interband velocity (off-diagonal terms in Hamiltonian gauge)

Transform the velocities into a gauge according to the following values of G

Wannier: keeps H, vs in the original, orbital basis

Hamiltonian: diagonalizes H and rotates H, vs into the energy, band basis

AutoBZ.get_safe_fermi_window_limits — Function

get_safe_fermi_window_limits(Ω, β, lb, ub)

Given a frequency, Ω, inverse temperature, β, returns an interval (l,u) with possibly truncated limits of integration for the frequency integral at each (Ω, β) point that are determined by the fermi_window_limits routine set to the default tolerances for the decay of the Fermi window function. The arguments lb and ub are lower and upper limits on the frequency to which the default result gets truncated if the default result would recommend a wider interval. If there is any truncation, a warning is emitted to the user, but the program will continue with the truncated limits.

Base.parent — Function

parent(::AbstractHamiltonianInterp)::FourierSeries

Return the Fourier series that the Hamiltonian wraps

parent(::AbstractVelocityInterp)::DerivativeSeries

Return the Hamiltonian object used for Wannier interpolation

AutoBZ.commutator — Function

commutator(A, B)

Return the commutator [A, B] = A*B - B*A.

AutoBZ.trgloc_integrand — Function

trgloc_integrand(h, M)

Returns tr(inv(M-h)) where M = ω*I-Σ(ω)

AutoBZ.covariant_velocity — Function

covariant_velocity(H, Hα, Aα)

Evaluates the velocity operator $\hat{v}_{\alpha} = -\frac{i}{\hbar} [\hat{r}_{\alpha}, \hat{H}]$ with the following expression, equivalent to eqn. 18 in Yates et al.

\[\hat{v}_{\alpha} = \frac{1}{\hbar} \hat{H}_{\alpha} + \frac{i}{\hbar} [\hat{H}, \hat{A}_{\alpha}]\]

where the $\alpha$ index implies differentiation by $k_{\alpha}$. Note that the terms that correct the derivative of the band velocity Also, this function takes $\hbar = 1$.

AutoBZ.coord — Function

coord(::AbstractCoordInterp{B})::AbstractCoordinate where B = B

Return the AbstractCoordinate basis in which an AbstractCoordInterp will be evaluated.

AutoBZ.LatticeRep — Type

LatticeRep()

Symmetry representation of objects that transform under the group action in the same way as the lattice.

AutoBZ.SOC — Type

SOC(A)

Wrapper for a matrix A that should be used as a block-diagonal matrix [A 0 0 A] as when dealing with spin-orbit coupling

AutoBZ.GaugeDefault — Function

GaugeDefault(::Type{T})::AbstractCoordinate where T

AbstractGaugeInterps should define this trait to declare the gauge that they assume their data is in.

AutoBZ.VcompDefault — Function

VcompDefault(::Type{T})::AbstractVelocityComponent where T

AbstractVelocityInterps should define this trait to declare the velocity component that they assume their data is in.

AutoBZ.herm — Function

herm(A::AbstractMatrix)

Return the Hermitian part of the matrix A, i.e. (A+A')/2.

AutoBZ.Lattice — Type

Lattice <: AbstractCoordinate

Singleton type representing lattice coordinates.